% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.ate.R
\name{as.data.table.ate}
\alias{as.data.table.ate}
\title{Turn ate Object Into a \code{data.table}}
\usage{
\method{as.data.table}{ate}(
  x,
  estimator = x$estimator,
  type = c("risk", "difference", "ratio"),
  keep.rownames = FALSE,
  ...
)
}
\arguments{
\item{x}{object obtained with function \code{ate}}

\item{estimator}{[character] The type of estimator relative to which the estimates should be output.}

\item{type}{[character vector] The type of risk to export.
Can be \code{"risk"} to export the risks specific to each treatment group,
\code{"difference"} to export the difference in risks between treatment groups,
or \code{"ratio"} to export the ratio of risks between treatment groups,.}

\item{keep.rownames}{Not used.}

\item{...}{Not used.}
}
\description{
Turn ate object into a \code{data.table}.
}
