% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.influenceTest.R
\name{as.data.table.influenceTest}
\alias{as.data.table.influenceTest}
\title{Turn influenceTest Object Into a \code{data.table}}
\usage{
\method{as.data.table}{influenceTest}(x, keep.rownames = FALSE, se = TRUE, ...)
}
\arguments{
\item{x}{object obtained with function \code{influenceTest}}

\item{keep.rownames}{Not used.}

\item{se}{[logical] Should standard errors/quantile for confidence bands be displayed?}

\item{...}{Not used.}
}
\description{
Turn influenceTest object into a \code{data.table}.
}
