% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.predictCSC.R
\name{as.data.table.predictCSC}
\alias{as.data.table.predictCSC}
\title{Turn predictCSC object into a \code{data.table}}
\usage{
\method{as.data.table}{predictCSC}(x, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{object obtained with function \code{predictCSC}}

\item{keep.rownames}{not used}

\item{...}{not used}
}
\description{
Turn predictCSC object into a \code{data.table}
}

