% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateCSC.R
\name{ateCSC}
\alias{ateCSC}
\title{Compute the average treatment effect using CSC}
\usage{
ateCSC(formula, data, treatment, contrasts = NULL, times, cause,
  n.bootstrap = 0, fitter = "cph", handler = c("mclapply",
  "sfClusterApplyLB"), mc.cores = 1, conf.level = 0.95,
  return.model = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{formula}{formula argument for CSC}

\item{data}{data argument for CSC}

\item{treatment}{name of the column containing the treatment
variable}

\item{contrasts}{the levels of treatment variable to be compared}

\item{times}{time points at which to evaluate risks}

\item{cause}{cause of intererst}

\item{n.bootstrap}{the number of bootstrap replications used to
compute the confidence interval}

\item{fitter}{Either \code{"cph"} or \code{"coxph"} passed to CSC}

\item{handler}{parallel handler for bootstrap. Either "mclapply" or
"sfClusterApplyLB"}

\item{mc.cores}{Passed on to \code{mclapply} or
\code{sfClusterApplyLB}. The number of cores to use, i.e. at
most how many child processes will be run simultaneously.  The
option is initialized from environment variable MC_CORES if
set.}

\item{conf.level}{the level for bootstrap confidence intervals}

\item{return.model}{Logical. if \code{TRUE} the fitted CSC is
returned as part of the output.}

\item{verbose}{Logical. If \code{TRUE} inform about estimated run time.}

\item{...}{passed to CSC}
}
\value{
A list with: point estimates, bootstrap quantile confidence intervals
model: the CSC model (optional)
}
\description{
Use the g-formula to estimate the average treatment effect
}
\examples{
dt <- SimCompRisk(1e3)
dt$time <- round(dt$time,1)
seqtimes <- sample(x = unique(sort(dt$time)), size = 100) 
dt$X1 <- factor(rbinom(1e3, prob = c(0.2,0.3,0.2) , size = 3), labels = paste0("T",0:3))
ateCSC(formula = Hist(time,event)~ X1+X2,data = dt, treatment = "X1", contrasts = NULL,
        times = 7, cause = 1, n.bootstrap = 3, mc.cores=1)

ateCSC(formula = Hist(time,event)~ X1+X2,data = dt, treatment = "X1", contrasts = NULL,
        times = 7, cause = 1, n.bootstrap = 3, mc.cores=2,fitter="cph")
}

