% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.ate.R, R/plot.R
\name{autoplot.ate}
\alias{autoplot.ate}
\alias{plot.ate}
\title{Plot Average Risks}
\usage{
\method{autoplot}{ate}(
  object,
  type = "meanRisk",
  first.derivative = FALSE,
  estimator = object$estimator[1],
  ci = object$inference$ci,
  band = object$inference$band,
  plot.type = "1",
  smooth = FALSE,
  digits = 2,
  alpha = NA,
  ylab = NULL,
  ...
)

\method{plot}{ate}(x, ...)
}
\arguments{
\item{object, x}{Object obtained with the function \code{ate}.}

\item{type}{[character vector] what to displayed.
Can be \code{"meanRisk"} to display the risks specific to each treatment group,
\code{"diffRisk"} to display the difference in risks between treatment groups,
\code{"ratioRisk"} to display the ratio of risks between treatment groups,
\code{"IPTW"} to display the treatment weights (if any) or \code{"probaT"} to display for each observation the probability of being assigned each treatment level,
and \code{"IPCW"} to display the censoring weights (if any).}

\item{first.derivative}{[logical] If \code{TRUE}, display the first derivative over time of the risks/risk differences/risk ratios.
(confidence intervals are obtained via simulation).}

\item{estimator}{[character] The type of estimator relative to which the risks should be displayed.}

\item{ci}{[logical] If \code{TRUE} display the confidence intervals for the average risks.}

\item{band}{[logical] If \code{TRUE} display the confidence bands for the average risks.}

\item{plot.type}{[character] Type of plot to be used.
\code{plot.type="2"} is useful when looking simulateneous at all eventtimes.
Otherwise use \code{plot.type="1"}.}

\item{smooth}{[logical] Should a smooth version of the risk function be plotted instead of a simple function?}

\item{digits}{[integer, >0] Number of decimal places.}

\item{alpha}{[numeric, 0-1] Transparency of the confidence bands. Argument passed to \code{ggplot2::geom_ribbon}.}

\item{ylab}{[character] Label for the y axis.}

\item{...}{Additional parameters to cutomize the display.}
}
\value{
A (possibly invisible) list containing:
\itemize{
\item plot: the ggplot object.
\item data: the data used to create the plot.
}
}
\description{
Plot average risks.
The method \code{plot} is a warper around \code{autoplot} that automatically produces a graphical display
but hides the corresponding dataset and ggplot2 object using \code{invisible}.
}
\section{Functions}{
\itemize{
\item \code{plot(ate)}: Graphical Display of Average Risks

}}
\examples{
library(survival)
library(ggplot2)

#### simulate data ####
n <- 1e2
set.seed(10)
dtS <- sampleData(n,outcome="survival")
seqTimes <- c(0,sort(dtS$time[dtS$event==1]),max(dtS$time))

#### Cox model ####
fit <- coxph(formula = Surv(time,event)~ X1+X2,data=dtS,y=TRUE,x=TRUE)

#### plot.type = 1: for few timepoints ####
ateFit <- ate(fit, data = dtS, treatment = "X1",
              times = c(1,2,5,10), se = TRUE, band = TRUE)
plot(ateFit)
\dontrun{
plot(ateFit, band = FALSE)
plot(ateFit, type = "diffRisk")
plot(ateFit, type = "ratioRisk")
}

#### plot.type = 2: when looking at all jump times ####
\dontrun{
ateFit <- ate(fit, data = dtS, treatment = "X1",
              times = seqTimes, se = TRUE, band = TRUE)

plot(ateFit)
plot(ateFit, plot.type = "2")

## customize plot
outGG <- autoplot(ateFit, plot.type = "2", alpha = 0.25)
outGG$plot + facet_wrap(~X1, labeller = label_both)


## Looking at the difference after smoothing
outGGS <-autoplot(ateFit, plot.type = "2", alpha = NA, smooth = TRUE)
outGGS$plot + facet_wrap(~X1, labeller = label_both)

## first derivative
## (computation of the confidence intervals takes time)
## (based on simulation - n.sim parameter)
autoplot(ateFit, plot.type = "2", smooth = TRUE,
         band = FALSE, type = "diffRisk")
autoplot(ateFit, plot.type = "2", smooth = TRUE, first.derivative = TRUE,
         band = FALSE, type = "diffRisk")
}
}
\seealso{
\code{\link{ate}} to compute average risks.
}
