% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.predictCox.R, R/plot.R
\name{autoplot.predictCox}
\alias{autoplot.predictCox}
\alias{plot.predictCox}
\title{Plot Predictions From a Cox Model}
\usage{
\method{autoplot}{predictCox}(
  object,
  type = NULL,
  ci = object$se,
  band = object$band,
  smooth = NULL,
  digits = 2,
  alpha = NA,
  group.by = NULL,
  reduce.data = FALSE,
  ylab = NULL,
  first.derivative = FALSE,
  atRisk = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{predictCox}(x, ...)
}
\arguments{
\item{object, x}{Object obtained with the function \code{predictCox}.}

\item{type}{[character] The type of predicted value to display.
Choices are:
\code{"hazard"} the hazard function,
\code{"cumhazard"} the cumulative hazard function, 
or \code{"survival"} the survival function.}

\item{ci}{[logical] If \code{TRUE} display the confidence intervals for the predictions.}

\item{band}{[logical] If \code{TRUE} display the confidence bands for the predictions.}

\item{smooth}{[logical] Should a smooth version of the risk function be plotted instead of a simple function?}

\item{digits}{[integer] Number of decimal places when displaying the values of the covariates in the caption.}

\item{alpha}{[numeric, 0-1] Transparency of the confidence bands. Argument passed to \code{ggplot2::geom_ribbon}.}

\item{group.by}{[character] The grouping factor used to color the prediction curves. Can be \code{"row"}, \code{"strata"}, or \code{"covariates"}.}

\item{reduce.data}{[logical] If \code{TRUE} only the covariates that does take indentical values for all observations are displayed.}

\item{ylab}{[character] Label for the y axis.}

\item{first.derivative}{[logical] If \code{TRUE}, display the first derivative over time of the risks/risk differences/risk ratios.
(confidence intervals are obtained via simulation).}

\item{atRisk}{[logical] Should the number at risk be diplayed at the bottom of the plot?}

\item{xlim, ylim}{[numeric vector of length 2] limits for the x and y axes.}

\item{...}{Additional parameters to cutomize the display: \itemize{
\item size.estimate: thickness of the line used to represent the estimated survival/hazard.
\item size.point: size of the points used to represent obsevations (if any).
\item size.ci: thickness of the line used to represent the confidence intervals for the survival/hazard.
\item size.band: thickness of the line used to represent the confidence bands for the survival/hazard.
\item size.atRisk: font size used to display the number at risk.
\item space.atRisk: vertical space used between the  display of the number at risk across strata. 
\item shape.point: shape of the points used to represent obsevations (if any).
\item col.ribbon: should shaded areas be colored according to the grouping factor. Otherwise grey. 
}}
}
\value{
A (possibly invisible) list containing:
\itemize{
\item plot: the ggplot object.
\item data: the data used to create the plot.
}
}
\description{
Plot predictions from a Cox model.
The method \code{plot} is a warper around \code{autoplot} that automatically produces a graphical display
but hides the corresponding dataset and ggplot2 object using \code{invisible}.
}
\section{Functions}{
\itemize{
\item \code{plot(predictCox)}: Graphical Display of Predictions From a Cox Model

}}
\examples{
library(survival)
library(ggplot2)

#### simulate data ####
set.seed(10)
d <- sampleData(1e2, outcome = "survival")
seqTau <- c(0,sort(unique(d$time[d$event==1])), max(d$time))

#### Cox model ####
m.cox <- coxph(Surv(time,event)~ X1 + X2 + X3,
                data = d, x = TRUE, y = TRUE)

## display baseline hazard
e.basehaz <- predictCox(m.cox)
plot(e.basehaz, type = "cumhazard")
\dontrun{
plot(e.basehaz, type = "cumhazard", size.point = 0) ## without points
plot(e.basehaz, type = "cumhazard", smooth = TRUE)
plot(e.basehaz, type = "cumhazard", smooth = TRUE, first.derivative = TRUE)
}

## display predicted survival
\dontrun{
pred.cox <- predictCox(m.cox, newdata = d[1:2,],
  times = seqTau, type = "survival", keep.newdata = TRUE)
plot(pred.cox)
plot(pred.cox, smooth = TRUE)
plot(pred.cox, group.by = "covariates")
plot(pred.cox, group.by = "covariates", reduce.data = TRUE)
plot(pred.cox, group.by = "X1", reduce.data = TRUE)
}

## predictions with confidence interval/bands
\dontrun{
pred.cox <- predictCox(m.cox, newdata = d[1:2,,drop=FALSE],
  times = seqTau, type = "survival", band = TRUE, se = TRUE, keep.newdata = TRUE)
res <- plot(pred.cox, ci = TRUE, band = TRUE)
res$plot + facet_wrap(~row)
res2 <- plot(pred.cox, ci = TRUE, band = TRUE, alpha = NA)
res2$plot + facet_wrap(~row, labeller = label_both) + guides(color = "none")
}

#### Stratified Cox model ####
\dontrun{
m.cox.strata <- coxph(Surv(time,event)~ strata(X1) + strata(X2) + X3 + X4,
                      data = d, x = TRUE, y = TRUE)

## baseline hazard
pred.baseline <- predictCox(m.cox.strata, keep.newdata = TRUE, type = "survival")
res <- plot(pred.baseline)
res$plot + facet_wrap(~strata, labeller = label_both)

## predictions
pred.cox.strata <- predictCox(m.cox.strata, newdata = d[1:3,,drop=FALSE],
                              time = seqTau, keep.newdata = TRUE, se = TRUE)

res2 <- plot(pred.cox.strata, type = "survival", group.by = "strata")
res2$plot + facet_wrap(~strata, labeller = label_both) + theme(legend.position="bottom")

## smooth version
plot(pred.cox.strata, type = "survival", group.by = "strata", smooth = TRUE, ci = FALSE)
}

#### Cox model with splines ####
\dontrun{
require(splines)
m.cox.spline <- coxph(Surv(time,event)~ X1 + X2 + ns(X6,4),
                data = d, x = TRUE, y = TRUE)
grid <- data.frame(X1 = factor(0,0:1), X2 = factor(0,0:1),
                   X6 = seq(min(d$X6),max(d$X6), length.out = 100))
pred.spline <- predictCox(m.cox.spline, newdata = grid, keep.newdata = TRUE,
                          se = TRUE, band = TRUE, centered = TRUE, type = "lp")
plot(pred.spline, group.by = "X6")
plot(pred.spline, group.by = "X6", alpha = 0.5, col.ribbon = TRUE)

grid2 <- data.frame(X1 = factor(1,0:1), X2 = factor(0,0:1),
                    X6 = seq(min(d$X6),max(d$X6), length.out = 100))
pred.spline <- predictCox(m.cox.spline, newdata = rbind(grid,grid2), keep.newdata = TRUE,
                          se = TRUE, band = TRUE, centered = TRUE, type = "lp")
autoplot(pred.spline, group.by = c("X6","X1"), alpha = 0.5)$plot + facet_wrap(~X1)
}
}
\seealso{
\code{\link{predictCox}} to compute cumulative hazard and survival based on a Cox model.
}
