% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.predictCSC.R, R/plot.R
\name{autoplot.predictCSC}
\alias{autoplot.predictCSC}
\alias{plot.predictCSC}
\title{Plot Predictions From a Cause-specific Cox Proportional Hazard Regression}
\usage{
\method{autoplot}{predictCSC}(
  object,
  ci = object$se,
  band = object$band,
  smooth = FALSE,
  digits = 2,
  alpha = NA,
  group.by = "row",
  reduce.data = FALSE,
  atRisk = FALSE,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{predictCSC}(x, ...)
}
\arguments{
\item{object, x}{Object obtained with the function \code{predictCox}.}

\item{ci}{[logical] If \code{TRUE} display the confidence intervals for the predictions.}

\item{band}{[logical] If \code{TRUE} display the confidence bands for the predictions.}

\item{smooth}{[logical] Should a smooth version of the risk function be plotted instead of a simple function?}

\item{digits}{[integer] Number of decimal places.}

\item{alpha}{[numeric, 0-1] Transparency of the confidence bands. Argument passed to \code{ggplot2::geom_ribbon}.}

\item{group.by}{[character] The grouping factor used to color the prediction curves. Can be \code{"row"}, \code{"strata"}, or \code{"covariates"}.}

\item{reduce.data}{[logical] If \code{TRUE} only the covariates that does take indentical values for all observations are displayed.}

\item{atRisk}{[logical] Should the number at risk be diplayed at the bottom of the plot?}

\item{xlim, ylim}{[numeric vector of length 2] limits for the x and y axes.}

\item{...}{Additional parameters to cutomize the display: \itemize{
\item size.estimate: thickness of the line used to represent the estimated survival/hazard.
\item size.point: size of the points used to represent obsevations (if any).
\item size.ci: thickness of the line used to represent the confidence intervals for the survival/hazard.
\item size.band: thickness of the line used to represent the confidence bands for the survival/hazard.
\item size.atRisk: font size used to display the number at risk.
\item space.atRisk: vertical space used between the  display of the number at risk across strata. 
\item shape.point: shape of the points used to represent obsevations (if any).
\item col.ribbon: should shaded areas be colored according to the grouping factor. Otherwise grey. 
}}
}
\value{
A (possibly invisible) list containing:
\itemize{
\item plot: the ggplot object.
\item data: the data used to create the plot.
}
}
\description{
Plot predictions from a Cause-specific Cox proportional hazard regression.
The method \code{plot} is a warper around \code{autoplot} that automatically produces a graphical display
but hides the corresponding dataset and ggplot2 object using \code{invisible}.
}
\section{Functions}{
\itemize{
\item \code{plot(predictCSC)}: Graphical Display of Predictions From a Cause-Specific Cox Model

}}
\examples{
library(survival)
library(rms)
library(ggplot2)
library(prodlim)

#### simulate data ####
set.seed(10)
d <- sampleData(1e2, outcome = "competing.risks")
seqTau <- c(0,unique(sort(d[d$event==1,time])), max(d$time))

#### CSC model ####
m.CSC <- CSC(Hist(time,event)~ X1 + X2 + X6, data = d)
pred.CSC <- predict(m.CSC, newdata = d[1:2,], time = seqTau, cause = 1, band = TRUE)
plot(pred.CSC, alpha = 0.2)

#### stratified CSC model ####
m.SCSC <- CSC(Hist(time,event)~ strata(X1) + strata(X2) + X6, data = d)
pred.SCSC <- predict(m.SCSC, time = seqTau, newdata = d[1:4,],
                     cause = 1, keep.newdata = TRUE, keep.strata = TRUE)
plot(pred.SCSC, group.by = "strata")
}
\seealso{
\code{\link{predict.CauseSpecificCox}} to compute risks based on a CSC model.
}
