% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{calcIChazard}
\alias{calcIChazard}
\title{Evaluate the influence function for the hazard functional}
\usage{
calcIChazard(ICcumhazard)
}
\arguments{
\item{ICcumhazard}{influence function of the cumulative hazard}
}
\value{
An object with the same dimensions as ICcumhazard, i.e. a list with one element per strata each element being a matrix observation*time.
}
\description{
Evaluate the influence function for the hazard functional using the one of the cumulative hazard
}
\examples{
\dontrun{
library(survival)

set.seed(10)
d <- SimSurv(1e2)
fit <- coxph(Surv(time,status)~X1 * X2,data=d, ties="breslow", x = TRUE, y = TRUE)

IFall <- iidCox(fit)
IFhaz <-  calcIChazard(IFall$ICLambda0)

cbind(IFhaz[[1]][,3], IFall$ICLambda0[[1]][,3]-IFall$ICLambda0[[1]][,2])
 
}
}
\author{
Brice Ozenne broz@sund.ku.dk
}

