% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{centerData}
\alias{centerData}
\alias{centerData.coxph}
\alias{centerData.cph}
\title{Center a dataset}
\usage{
centerData(object, data, stratavars)

\method{centerData}{cph}(object, data = model.frame(object), stratavars,
  center)

\method{centerData}{coxph}(object, data = model.frame(object), stratavars,
  center)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{data}{a \code{data.frame} or a \code{data.table}}

\item{stratavars}{the name of the strata variables}

\item{center}{should the covariates be centered}
}
\description{
Center a dataset around the centering values
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCoxS <- coxph(Surv(time, event) ~ X1, data = d) 
centerData(mCoxS, stratavars = NULL)
mCoxS <- coxph(Surv(time, event) ~ X1+X2, data = d) 
centerData(mCoxS, stratavars = NULL)
mCoxS <- coxph(Surv(time, event) ~ strata(X1)+X2, data = d) 
centerData(mCoxS, stratavars = "strata(X1)")
mCoxS <- coxph(Surv(time, event) ~ strata(X1)+strata(X2), data = d) 
centerData(mCoxS, stratavars = c("strata(X1)","strata(X2)"))

##
library(rms)
mCoxS <- cph(Surv(time, event) ~ X1, data = d, y = TRUE) 
centerData(mCoxS, stratavars = NULL)
mCoxS <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE) 
centerData(mCoxS, stratavars = NULL)
mCoxS <- cph(Surv(time, event) ~ strat(X1)+X2, data = d, y = TRUE) 
centerData(mCoxS, stratavars = "strat(X1)")
mCoxS <- cph(Surv(time, event) ~ strat(X1)+strat(X2), data = d, y = TRUE) 
centerData(mCoxS, stratavars = c("strat(X1)","strat(X2)"))
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

