% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.party.R
\name{Cforest}
\alias{Cforest}
\title{S3-wrapper function for cforest from the party package}
\usage{
Cforest(formula, data, ...)
}
\arguments{
\item{formula}{Passed on as is. See \code{cforest} of the \code{party} package}

\item{data}{Passed on as is. See \code{cforest} of the \code{party} package}

\item{...}{Passed on as they are. See \code{cforest} of the \code{party} package}
}
\value{
list with two elements: cforest and call
}
\description{
S3-wrapper function for cforest from the party package
}
\details{
See \code{cforest} of the \code{party} package.
}
\references{
Ulla B. Mogensen, Hemant Ishwaran, Thomas A. Gerds (2012).
Evaluating Random Forests for Survival Analysis Using Prediction Error
Curves. Journal of Statistical Software, 50(11), 1-23. URL
http://www.jstatsoft.org/v50/i11/.
}
\keyword{survival}
