% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.riskRegression.R
\name{coef.riskRegression}
\alias{coef.riskRegression}
\title{Extract coefficients from riskRegression model}
\usage{
\method{coef}{riskRegression}(object, digits = 3, eps = 10^-4, ...)
}
\arguments{
\item{object}{Object obtained with \code{ARR} or \code{LRR} or \code{riskRegression}}

\item{digits}{Number of digits}

\item{eps}{P-values below this number are shown as \code{<eps}}

\item{...}{not used}
}
\description{
Extract coefficients from riskRegression model
}
