% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{coef.wglm}
\alias{coef.wglm}
\title{Estimates from IPCW Logistic Regressions}
\usage{
\method{coef}{wglm}(object, times = NULL, simplify = TRUE, ...)
}
\arguments{
\item{times}{[numeric vector] time points at which the estimates should be output.}

\item{simplify}{[logical] should the ouput be converted to a vector when only one timepoint is requested. Otherwise will always return a matrix.}

\item{...}{Not used.}

\item{x}{a wglm object.}
}
\description{
Display the estimated regression parameters from logistic regressions.
}
