% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colCumProd}
\alias{colCumProd}
\title{Apply cumprod in each column}
\usage{
colCumProd(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of apply(x,2,cumprod)
}
\examples{
x <- matrix(1:8,ncol=2)
colCumProd(x)
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}
