% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colCumSum}
\alias{colCumSum}
\title{Apply cumsum in each column}
\usage{
colCumSum(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of apply(x,2,cumsum)
}
\examples{
x <- matrix(1:8,ncol=2)
colCumSum(x)
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}

