% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confBandCox.R
\name{confBandCox}
\alias{confBandCox}
\title{Compute quantiles of a gaussian process}
\usage{
confBandCox(iid, se, times, n.sim, conf.level)
}
\arguments{
\item{iid}{The iid decomposition of the estimator over time.}

\item{se}{The variance of the estimate over time.}

\item{times}{The time points covered by the confidence bands.}

\item{n.sim}{The number of simulations used to compute the quantiles.}

\item{conf.level}{Level of confidence.}
}
\description{
Compute quantiles of a gaussian process
}

