% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.ate.R
\name{confint.ate}
\alias{confint.ate}
\title{Confidence Intervals and Confidence Bands for the Predicted Absolute Risk (Cumulative Incidence Function)}
\usage{
\method{confint}{ate}(object, conf.level = 0.95, nsim.band = 10000,
  ate.transform = "none", diffAte.transform = "none",
  ratioAte.transform = "none", seed = NA, type.boot = "perc", ...)
}
\arguments{
\item{object}{A \code{ate} object, i.e. output of the \code{ate} function.}

\item{conf.level}{Level of confidence.}

\item{nsim.band}{the number of simulations used to compute the quantiles for the confidence bands.}

\item{seed}{Integer passed to set.seed when performing simulation for the confidence bands.
If not given or NA no seed is set.}

\item{...}{not used.}

\item{transform.absRisk}{the transformation used to improve coverage
of the confidence intervals for the predicted absolute risk in small samples.}
}
\description{
Confidence intervals and confidence Bands for the predicted absolute risk (cumulative incidence function).
}
\details{
The confidence bands and confidence intervals are automatically restricted to the interval [0;1].
}
\examples{
library(survival)

## generate data
set.seed(10)
d <- sampleData(40,outcome="survival") ## training dataset
nd <- sampleData(4,outcome="survival") ## validation dataset
d$time <- round(d$time,1) ## create tied events
# table(duplicated(d$time))

## estimate a stratified Cox model
fit <- coxph(Surv(time,event)~X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)


## compute individual specific cumulative hazard and survival probabilities 
fit.pred <- predictCox(fit, newdata=nd, times=c(3,8), se = TRUE)
fit.pred

## add confidence intervals computed on the original scale
confint(fit.pred, transform.survival = "none")
fit.pred$survival + 1.96 * fit.pred$survival.se  ## survival.upper

## add confidence intervals computed on the log-log scale
## and backtransformed
confint(fit.pred, transform.survival = "loglog")
newse <- fit.pred$survival.se/(fit.pred$survival*log(fit.pred$survival))
exp(-exp(log(-log(fit.pred$survival)) - 1.96 * newse)) ## survival.lower
exp(-exp(log(-log(fit.pred$survival)) + 1.96 * newse)) ## survival.upper
}
\author{
Brice Ozenne
}
