% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.ateRobust.R
\name{confint.ateRobust}
\alias{confint.ateRobust}
\title{Confidence Intervals for the Average Treatment Effect}
\usage{
\method{confint}{ateRobust}(object, parm = NULL, level = 0.95,
  meanRisk.transform = "none", diffRisk.transform = "none", ...)
}
\arguments{
\item{object}{A \code{ateRobust} object, i.e. output of the \code{ateRobust} function.}

\item{parm}{not used. For compatibility with the generic method.}

\item{level}{[numeric, 0-1] Level of confidence.}

\item{meanRisk.transform}{[character] the transformation used
to improve coverage of the confidence intervals for the mean risk in small samples.
Can be \code{"none"}, \code{"log"}, \code{"loglog"}, \code{"cloglog"}.}

\item{diffRisk.transform}{[character] the transformation used to improve coverage
of the confidence intervals for the risk difference in small samples.
Can be \code{"none"}, \code{"atanh"}.}

\item{...}{not used.}
}
\description{
Confidence intervals for the average treatment effect
}
\author{
Brice Ozenne
}
