% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.influenceTest.R
\name{confint.influenceTest}
\alias{confint.influenceTest}
\title{Confidence Intervals and Confidence Bands for the Difference Between Two Estimates}
\usage{
\method{confint}{influenceTest}(object, parm = NULL, level = 0.95,
  nsim.band = 10000, transform = "none", seed = NA, ...)
}
\arguments{
\item{object}{A \code{influenceTest} object, i.e. output of the \code{influenceTest} function.}

\item{parm}{not used. For compatibility with the generic method.}

\item{level}{[numeric, 0-1] Level of confidence.}

\item{nsim.band}{[integer, >0] the number of simulations used to compute the quantiles for the confidence bands.}

\item{transform}{[character] the transformation used to improve coverage of the confidence intervals.
Can be \code{"none"} or \code{"atanh"}.}

\item{seed}{[integer, >0] seed number set before performing simulations for the confidence bands.
If not given or NA no seed is set.}

\item{...}{not used.}
}
\description{
Confidence intervals and confidence Bands for the difference between two estimates.
}
\details{
Except for the cumulative hazard,
the confidence bands and confidence intervals are automatically restricted to the interval [-1;1].
}
\author{
Brice Ozenne
}
