% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.predictCox.R
\name{confint.predictCox}
\alias{confint.predictCox}
\title{Confidence Intervals and Confidence Bands for the predicted Survival/Cumulative Hazard}
\usage{
\method{confint}{predictCox}(object, conf.level = 0.95, type = NULL,
  nsim.band = 10000, cumhazard.transform = "log",
  survival.transform = "cloglog", seed = NA, ...)
}
\arguments{
\item{object}{A \code{predictCox} object, i.e. output of the \code{predictCox} function.}

\item{conf.level}{Level of confidence.}

\item{type}{the type of predicted value for which the confidence intervals should be output.
Can be \code{"survival"} or \code{"cumhazard"}.}

\item{nsim.band}{the number of simulations used to compute the quantiles for the confidence bands.}

\item{cumhazard.transform}{the transformation used to improve coverage
of the confidence intervals for the cumlative hazard in small samples.}

\item{survival.transform}{the transformation used to improve coverage
of the confidence intervals for the survival in small samples.}

\item{seed}{Integer passed to set.seed when performing simulation for the confidence bands.
If not given or NA no seed is set.}

\item{...}{not used.}
}
\description{
Confidence intervals and confidence Bands for the predicted survival/cumulative Hazard.
}
\details{
The confidence bands and confidence intervals are automatically restricted to the interval of definition of the statistic,
i.e. a confidence interval for the survival of [0.5;1.2] will become [0.5;1].
}
\examples{
library(survival)

## generate data
set.seed(10)
d <- sampleData(40,outcome="survival") ## training dataset
nd <- sampleData(4,outcome="survival") ## validation dataset
d$time <- round(d$time,1) ## create tied events
# table(duplicated(d$time))

## estimate a stratified Cox model
fit <- coxph(Surv(time,event)~X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)


## compute individual specific cumulative hazard and survival probabilities 
fit.pred <- predictCox(fit, newdata=nd, times=c(3,8), se = TRUE)
fit.pred

## add confidence intervals computed on the original scale
confint(fit.pred, transform.survival = "none")
fit.pred$survival + 1.96 * fit.pred$survival.se  ## survival.upper

## add confidence intervals computed on the log-log scale
## and backtransformed
confint(fit.pred, transform.survival = "loglog")
newse <- fit.pred$survival.se/(fit.pred$survival*log(fit.pred$survival))
exp(-exp(log(-log(fit.pred$survival)) - 1.96 * newse)) ## survival.lower
exp(-exp(log(-log(fit.pred$survival)) + 1.96 * newse)) ## survival.upper
}
\author{
Brice Ozenne
}
