% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{confint.wglm}
\alias{confint.wglm}
\title{Confidence intervals of IPCW Logistic Regressions}
\usage{
\method{confint}{wglm}(
  object,
  parm,
  level = 0.95,
  se = "robust",
  times = NULL,
  simplifies = TRUE,
  ...
)
}
\arguments{
\item{parm}{not used.}

\item{level}{[numeric] level of the confidence intervals.}

\item{se}{[character] Should robust standard error be used, accounting for the uncertainty of the censoring weights (\code{"robust"}),
or ignoring the uncertainty of the censoring weights (\code{"robust-wknown"}), or model-based standard errors ignoring the uncertainty of the censoring weights (\code{"model-wknown"}).}

\item{times}{[numeric vector] time points at which the coefficients should be output.}

\item{simplifies}{[logical] should the ouput be converted to a data.frame when only one timepoint is requested. Otherwise will always return a list.}

\item{...}{Not used.}

\item{x}{a wglm object.}
}
\description{
Extract regression coefficients (relative to the outcome model) of a IPCW Logisitc model,
along with their confidence intervals and p-values.
}
