% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxBaseEstimator}
\alias{coxBaseEstimator}
\alias{coxBaseEstimator.coxph}
\alias{coxBaseEstimator.phreg}
\title{Extract the type of estimator for the baseline hazard}
\usage{
coxBaseEstimator(object)

\method{coxBaseEstimator}{coxph}(object)

\method{coxBaseEstimator}{phreg}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}
}
\description{
Extract the type of estimator for the baseline hazard
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")
d$entry <- 0

##
library(survival)
mcox <- coxph(Surv(time, event) ~ X1+X2, data = d)
coxBaseEstimator(mcox)

##
library(rms)
mcox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
coxBaseEstimator(mcox)

##
library(mets)
mcox <- phreg(Surv(entry, time, event) ~ X1+X2, data = d)
coxBaseEstimator(mcox)
}
}
\author{
Brice Ozenne broz@sund.ku.dk
}
