% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxDesign}
\alias{CoxDesign}
\title{Extract the design matrix for a Cox model}
\usage{
CoxDesign(object, data, lpvars, stratavars, rm.intercept, center)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{data}{a \code{data.frame} or a \code{data.table}}

\item{lpvars}{the variables used to define the linear predictor}

\item{stratavars}{the special variables used to define the strata}

\item{rm.intercept}{should the intercept be removed}

\item{center}{should the covariates be centered}
}
\description{
Extract the design matrix for a Cox model
}
\details{
if no variables for the linear predictor or the strata, returns a data.table with one column filled with 0.
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCoxS <- coxph(Surv(time, event) ~ X1, data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)
CoxDesign(mCoxS, data = d, lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = TRUE)

mCoxS <- coxph(Surv(time, event) ~ X1*X2, data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)

mCoxS <- coxph(Surv(time, event) ~ strata(X1)+X2, data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)

mCoxS <- coxph(Surv(time, event) ~ strata(X1)+strata(X2), data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)

##
library(rms)
mCoxS <- cph(Surv(time, event) ~ X1, data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = TRUE)

mCoxS <- cph(Surv(time, event) ~ X1*X2, data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)

mCoxS <- cph(Surv(time, event) ~ strat(X1)+X2, data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)

mCoxS <- cph(Surv(time, event) ~ strat(X1)+strat(X2), data = d) 
info <- CoxStrataVar(mCoxS)
CoxDesign(mCoxS, data = CoxData(mCoxS), lpvars = info$lpvars, stratavars = info$stratavars, rm.intercept = TRUE, center = FALSE)
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

