% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxEventtime}
\alias{CoxEventtime}
\alias{CoxEventtime.coxph}
\alias{CoxEventtime.cph}
\title{Extract the event times used to fit a Cox model}
\usage{
CoxEventtime(object)

\method{CoxEventtime}{cph}(object)

\method{CoxEventtime}{coxph}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}
}
\description{
Extract the event times used to fit a Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
CoxEventtime(mCox)

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
CoxEventtime(mCox)
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

