% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxN}
\alias{coxN}
\alias{coxN.cph}
\alias{coxN.coxph}
\alias{coxN.phreg}
\title{Extract the number of observations from a Cox model}
\usage{
coxN(object)

\method{coxN}{cph}(object)

\method{coxN}{coxph}(object)

\method{coxN}{phreg}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}
}
\description{
Extract the number of observations from a Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")
d$entry <- 0

##
library(survival)
mcox <- coxph(Surv(time, event) ~ X1+X2, data = d)
coxN(mcox)

##
library(rms)
mcox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
coxN(mcox)

##
library(mets)
mcox <- phreg(Surv(entry, time, event) ~ X1+X2, data = d)
coxN(mcox)
}
}
\author{
Brice Ozenne broz@sund.ku.dk
}
