% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxResponseVar}
\alias{CoxResponseVar}
\title{Extract the time and event variable from a Cox mode}
\usage{
CoxResponseVar(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}
}
\description{
Extract the time and event variable from a Cox mode
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
CoxResponseVar(mCox)

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
CoxResponseVar(mCox) 
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

