% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxSpecial.R
\name{coxSpecial}
\alias{coxSpecial}
\alias{coxSpecial.coxph}
\alias{coxSpecial.cph}
\alias{coxSpecial.phreg}
\alias{coxSpecial.prodlim}
\alias{coxSpecial.coxnet}
\title{Special characters in Cox model}
\usage{
coxSpecial(object)

\method{coxSpecial}{coxph}(object)

\method{coxSpecial}{cph}(object)

\method{coxSpecial}{phreg}(object)

\method{coxSpecial}{prodlim}(object)

\method{coxSpecial}{coxnet}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}
}
\description{
Return the special character(s) of the Cox model, e.g. used to indicate the strata variables.
}
\details{
Must return a list with at least one element strata
indicating the character in the formula marking the variable(s) defining the strata.
}
\author{
Brice Ozenne broz@sund.ku.dk
}
