% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxStatus}
\alias{CoxStatus}
\alias{CoxStatus.coxph}
\alias{CoxStatus.cph}
\title{Extract the event indicator used to fit a Cox model}
\usage{
CoxStatus(object)

\method{CoxStatus}{cph}(object)

\method{CoxStatus}{coxph}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}
}
\description{
Extract the event indicator used to fit a Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
CoxStatus(mCox)

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
CoxStatus(mCox)
}
}
\author{
Brice Ozenne broz@sund.ku.dk
}

