% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxStrataLevel}
\alias{coxStrataLevel}
\alias{coxStrataLevel.coxph}
\alias{coxStrataLevel.cph}
\alias{coxStrataLevel.phreg}
\alias{coxStrataLevel.prodlim}
\title{Returns the name of the strata in Cox model}
\usage{
coxStrataLevel(object)

\method{coxStrataLevel}{coxph}(object)

\method{coxStrataLevel}{cph}(object)

\method{coxStrataLevel}{phreg}(object)

\method{coxStrataLevel}{prodlim}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}
}
\description{
Return the name of the strata in Cox model
}
\author{
Brice Ozenne broz@sund.ku.dk
}
