% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxStrataVar}
\alias{CoxStrataVar}
\alias{CoxStrataVar.coxph}
\alias{CoxStrataVar.cph}
\title{Extract variable names from a model}
\usage{
CoxStrataVar(object)

\method{CoxStrataVar}{coxph}(object)

\method{CoxStrataVar}{cph}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}
}
\description{
Extract the name of the variables belonging to the linear predictor or used to form the strata
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
CoxStrataVar(mCox)
mCoxS <- coxph(Surv(time, event) ~ strata(X1)+X2, data = d)
CoxStrataVar(mCoxS)
mCoxS2 <- coxph(Surv(time, event) ~ strata(X1)+strata(X2), data = d)
CoxStrataVar(mCoxS2)
mCoxI <- coxph(Surv(time, event) ~ X1*X2, data = d)
CoxStrataVar(mCoxI)

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
CoxStrataVar(mCox) 
mCoxS <- cph(Surv(time, event) ~ strat(X1)+X2, data = d)
CoxStrataVar(mCoxS) 
mCoxS2 <- cph(Surv(time, event) ~ strat(X1)+strat(X2), data = d, y = TRUE)
CoxStrataVar(mCoxS2) 
mCoxI <- cph(Surv(time, event) ~ X1*X2, data = d)
CoxStrataVar(mCoxI)
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

