% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.party.R
\name{Ctree}
\alias{Ctree}
\title{S3-Wrapper for ctree.}
\usage{
Ctree(...)
}
\arguments{
\item{...}{passed to ctree}
}
\value{
list with two elements: ctree and call
}
\description{
The call is added to an ctree object
}
\examples{
if (requireNamespace("party",quietly=TRUE)){
library(prodlim)
library(party)
library(survival)
set.seed(50)
d <- SimSurv(50)
nd <- data.frame(X1=c(0,1,0),X2=c(-1,0,1))
f <- Ctree(Surv(time,status)~X1+X2,data=d)
predictRisk(f,newdata=nd,times=c(3,8))
}


}
\seealso{
Cforest
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
