% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{defineStrata}
\alias{defineStrata}
\alias{defineStrata.coxph}
\alias{defineStrata.cph}
\title{Define the strata for a new dataset}
\usage{
defineStrata(object, data, sterms, stratavars, levelsStrata, ...)

\method{defineStrata}{cph}(object, data, sterms, stratavars, levelsStrata, ...)

\method{defineStrata}{coxph}(object, data, sterms, stratavars, levelsStrata,
  stratalevels)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{data}{a \code{data.frame} or a \code{data.table}}

\item{sterms}{terms in the formula corresponding to the strata variables}

\item{stratavars}{the name of the variables used to define the strata}

\item{levelsStrata}{the strata levels that have been used to fit the Cox model}

\item{...}{additional arguments to be passed to the low level functions}

\item{stratalevels}{a named list containing for each variable used to form the strata all its possible levels}
}
\description{
Define the strata in a dataset to match those of a stratified Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCoxS <- coxph(Surv(time, event) ~ strata(X1)+X2, data = d)
info <- getCoxInfo(mCoxS, design = FALSE)
defineStrata(mCoxS, data = d, sterms = info$sterms, stratavars = info$stratavars, 
             levelsStrata = levels(info$strataF), stratalevels = info$stratalevels) 

##
library(rms)
mCoxS <- cph(Surv(time, event) ~ strat(X1)+X2, data = d, y = TRUE)
info <- getCoxInfo(mCoxS, design = FALSE)
defineStrata(mCoxS, data = d, sterms = info$sterms, stratavars = info$stratavars, 
             levelsStrata = levels(info$strataF), stratalevels = info$stratalevels) 
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

