% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPWbox.R
\name{drop1.IPWbox}
\alias{drop1.IPWbox}
\title{Update Large IPW values}
\usage{
\method{drop1}{IPWbox}(object, scope, method = "trim", normalize = TRUE, ...)
}
\arguments{
\item{object}{[IPWbox] output of IPWbox.}

\item{scope}{[numeric,>0] value beyond which the weight should be modified.}

\item{method}{[character] either set the weight to 0 (\code{"trim"}) by setting the corresponding probability to infinity
or set the weight to the value of argument scope (\code{"threshold"}).}

\item{normalize}{[logical] if \code{TRUE} will rescale the 'non-modified' weights (i.e. whose value is below scope)
to match the initial population size. The sum by columns of the weights is then unchanged between before and after the application of drop1.}

\item{...}{not used. For compatibility with the generic method}
}
\value{
An IPWbox object
}
\description{
Update IPW larger than a user-defined value, either setting them to 0 or to a fixed value.
}
\details{
WARNING: how to update the influence function is not well established and a pragmatic choice has been made: \itemize{
\item \code{"trim"}: influence function for the weights set to 0 is also set to 0.
\item \code{"threshold"}: influence function is unchanged
}
but there is approach has not be proven to control the type 1 error.
}
