% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{extractStrata}
\alias{extractStrata}
\title{Extract the information about the strata}
\usage{
extractStrata(xterms, xlevels = NULL, special)
}
\arguments{
\item{xterms}{...}

\item{xlevels}{...}

\item{special}{the special character indicating the strata variables}
}
\description{
Extract the information about the strata stored in a Cox model
}
\author{
Brice Ozenne broz@sund.ku.dk and Thomas A. Gerds tag@biostat.ku.dk
}
