% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{getCoxInfo}
\alias{getCoxInfo}
\alias{getCoxInfo.coxph}
\alias{getCoxInfo.cph}
\title{Extract information from a Cox model}
\usage{
getCoxInfo(object, design, center)

\method{getCoxInfo}{cph}(object, design, center)

\method{getCoxInfo}{coxph}(object, design, center)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{design}{should the design matrix be returned?}

\item{center}{should the covariates be centered}
}
\value{
A named list containing the following elements:
\itemize{
 \item{"nPatients"}{the number of observations}
 \item{"xvars"}{the name of all the regressors (including those used to form the strata)}
 \item{"modeldata"}{the design matrix for the regressors}
 \item{"stratavars"}{the name of the variables used to define the strata}
 \item{"is.strata"}{is there any strata?}
 \item{"sterms"}{terms corresponding to the strata variables}
 \item{"strataF"}{a vector contain the strata factor for each observation}
 \item{"stratalevels"}{a named list containing for each variable used to form the strata all its possible levels}
}
}
\description{
Extract information from a Cox model that are necessary for computing the baseline hazard, performing predictions or estimating the influence function associated to the Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
getCoxInfo(mCox, design = FALSE) 
mCoxS <- coxph(Surv(time, event) ~ strata(X1)+strata(X2), data = d)
getCoxInfo(mCoxS, design = FALSE) 

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
getCoxInfo(mCox, design = FALSE) 
mCoxS <- cph(Surv(time, event) ~ strat(X1)+strat(X2), data = d, y = TRUE)
getCoxInfo(mCoxS, design = FALSE) 
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

