% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAL9001.R
\name{Hal9001}
\alias{Hal9001}
\title{Fitting HAL for use with predictRisk}
\usage{
Hal9001(formula, data, lambda = NULL, ...)
}
\arguments{
\item{formula}{A formula.}

\item{data}{The data on which to fit the model.}

\item{lambda}{The tuning parameters for HAL. If set to NULL, then it the parameters are chosen for you.}

\item{\dots}{Additional arguments that are passed on to the hal_fit.}
}
\description{
Fit HAL models via a formula and a data set for use with \code{\link{predictRisk}}.
}
