% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{IClambda2hazard}
\alias{IClambda2hazard}
\title{Evaluate the influence function for the hazard based on the one of the baseline hazard##'}
\usage{
IClambda2hazard(eXb, X_ICbeta, lambda0, IClambda0)
}
\arguments{
\item{eXb}{the linear predictor}

\item{X_ICbeta}{the design matrix times the influence function of beta}

\item{lambda0}{the baseline hazard}

\item{IClambda0}{the influence function of the baseline hazard}
}
\description{
Evaluate the influence function for the hazard based on the one of the baseline hazard
}

