% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSeCox.R
\name{IFlambda2hazard}
\alias{IFlambda2hazard}
\title{Evaluate the influence function for the hazard based on the one of the baseline hazard##'}
\usage{
IFlambda2hazard(eXb, X_IFbeta, lambda0, IFlambda0, nVar)
}
\arguments{
\item{eXb}{the linear predictor}

\item{X_IFbeta}{the design matrix times the influence function of beta}

\item{lambda0}{the baseline hazard}

\item{IFlambda0}{the influence function of the baseline hazard}

\item{nVar}{the number of variables that form the linear predictor}
}
\description{
Evaluate the influence function for the hazard based on the one of the baseline hazard
}
