% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{iid.wglm}
\alias{iid.wglm}
\title{IID for IPCW Logistic Regressions}
\usage{
\method{iid}{wglm}(x, times = NULL, simplify = TRUE, store = NULL, ...)
}
\arguments{
\item{x}{a wglm object.}

\item{times}{[numeric vector] time points at which the iid should be output.}

\item{simplify}{[logical] should the ouput be converted to a matrix when only one timepoint is requested. Otherwise will always return a list.}

\item{store}{[vector] when evaluating the iid, should prediction be only computed for unique covariate sets and mapped back to the original dataset (\code{data="minimal"}). Otherwise use \code{data="full"}.}

\item{...}{Not used.}
}
\description{
Compute the decomposition in iid elements of the ML estimor of IPCW logistic regressions.
}
