% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{information.wglm}
\alias{information.wglm}
\title{Information for IPCW Logistic Regressions}
\usage{
\method{information}{wglm}(x, times = NULL, simplifies = TRUE, ...)
}
\arguments{
\item{x}{a wglm object.}

\item{times}{[numeric vector] time points at which the score should be output.}

\item{simplifies}{[logical] should the ouput be converted to a matrix when only one timepoint is requested. Otherwise will always return a list.}

\item{...}{Not used.}
}
\description{
Compute the information (i.e. opposit of the expectation of the second derivative of the log-likelihood) for IPCW logistic regressions.
}
