% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPWbox.R
\name{IPWbox}
\alias{IPWbox}
\title{Encapsulate Weights}
\usage{
IPWbox(variable, proba, IF = NULL, id = NULL)
}
\arguments{
\item{variable}{[character vector] name of the treatment variable or (time,censoring) variables .}

\item{proba}{[matrix] value of the weight for each observation (rows) and treatment group or timepoints (columns).}

\item{IF}{[array] value of the influence function for each observation (rows) relative to each weight (columns,slice).}

\item{id}{[list of a single vector ] optional id variable.}
}
\value{
An object
}
\description{
Object containing user-defined weights that can be used with the \code{ate} function.
}
\details{
Enable to specified user-defined treatment and censoring weights with the ate function.
For instance if one wants to truncate the weights or restrict the population to non-extreme weights.
}
\examples{

library(data.table)

#############################################
#### Survival settings without censoring ####
#### ATE with glm                        ####
#############################################

#### generate data ####
n <- 250
dt <- sampleData(n, outcome="binary")
dt[, X2 := as.numeric(X2)]

#### compute the ATE ####
m.event <- glm(formula = Y ~ X1+X2+X3, data=dt, family = "binomial")
m.treatment <- glm(X3~X1+X2+X5+X8, data=dt,family=binomial(link="logit"))
ateFit <- ate(m.event, treatment = m.treatment, data = dt, store = c(weights = TRUE))
model.tables(ateFit)

range(weights(ateFit, type = "IPTW"))

manualWeights <- weights(ateFit, type = "IPTWbox")
manualWeights$proba
ateFit.bis <- ate(m.event, treatment = manualWeights, data = dt)
model.tables(ateFit.bis) ## same as when providing the logistic model

#### set weights above 5 to 0
manualWeights.trim <- drop1(manualWeights, scope = 5)
ateFit.trim <- ate(m.event, treatment = manualWeights.trim, data = dt)
model.tables(ateFit.trim)

}
