% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iidCox.R
\name{is.iidCox}
\alias{is.iidCox}
\title{Check Computation of the Influence Function in a Cox Model}
\usage{
is.iidCox(object)
}
\arguments{
\item{object}{fitted Cox regression model object either obtained with \code{coxph} (survival package), \code{cph} (rms package), \code{CSC} (riskRegression package).}
}
\description{
Check whether the influence function of the Cox model or cause specific Cox models has been stored in the object.
}
