% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{lpCox}
\alias{lpCox}
\alias{lpCox.coxph}
\alias{lpCox.cph}
\title{Compute the linear predictor of a Cox model}
\usage{
lpCox(object, data, xvars, stratavars)

\method{lpCox}{cph}(object, data, xvars, stratavars)

\method{lpCox}{coxph}(object, data, xvars, stratavars)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{data}{a \code{data.frame} or a \code{data.table}}

\item{xvars}{the name of all the regressors (including those used to form the strata)}

\item{stratavars}{the variables used to define the strata}
}
\description{
Compute the linear predictor of a Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
lpCox(mCox, data = d, xvars = c("X1","X2"), stratavars = NULL) 
mCoxS <- coxph(Surv(time, event) ~ strata(X1)+X2, data = d)
lpCox(mCoxS, data = d, xvars = c("X1","X2"), stratavars = "X1") 
mCoxS2 <- coxph(Surv(time, event) ~ strata(X1)+strata(X2), data = d)
lpCox(mCoxS2, data = d, xvars = c("X1","X2"), stratavars = c("X1","X2")) 

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
lpCox(mCox, data = d, xvars = c("X1","X2"), stratavars = NULL) 
mCoxS <- cph(Surv(time, event) ~ strat(X1)+X2, data = d)
lpCox(mCoxS, data = d, xvars = c("X1","X2"), stratavars = "X1") 
mCoxS2 <- cph(Surv(time, event) ~ strat(X1)+strat(X2), data = d, y = TRUE)
lpCox(mCoxS2, data = d, xvars = c("X1","X2"), stratavars = c("X1","X2")) 
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

