% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{model.tables.wglm}
\alias{model.tables.wglm}
\title{Statistical Inference for Estimate from IPCW Logistic Regressions}
\usage{
\method{model.tables}{wglm}(x, times = NULL, type = "robust", level = 0.95, ...)
}
\arguments{
\item{x}{a wglm object.}

\item{times}{[numeric vector] time points at which the estimates should be output.}

\item{type}{[character] should the robust variance-covariance matrix be computing accounting for the uncertainty of the IPCW (\code{"robust"})
or ignoring the uncertainty of the IPCW (\code{"robust-wknown"}), or model-based ignoring the uncertainty of the IPCW  (\code{"model-wknown"})?}

\item{level}{[numeric, 0-1] Level of confidence.}

\item{...}{Not used.}
}
\description{
Export estimated regression parameters from IPCW logistic regressions with their uncertainty (standard errors, confidence intervals and p-values).
}
