% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScore.R
\name{plot.highscore}
\alias{plot.highscore}
\title{plot score object}
\usage{
\method{plot}{highscore}(x, models, xlab, ylab, xlim, ylim, col, lwd = 3, lty,
  type, add = FALSE, ...)
}
\arguments{
\item{x}{Risk prediction score}

\item{models}{Choice of models to plot}

\item{xlab}{label for x-axis}

\item{ylab}{label for y-axis}

\item{xlim}{limits for x-axis}

\item{ylim}{limits for x-axis}

\item{col}{line color}

\item{lwd}{line width}

\item{lty}{line type}

\item{type}{Either \code{"AUC"} or \code{"Brier"}}

\item{add}{Logical. If \code{TRUE} add to existing plot}

\item{...}{passed to plot}
}
\description{
plot score object
}

