% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predictCox.R
\name{plot.predictCox}
\alias{plot.predictCox}
\title{Plot predictions from a Cox model}
\usage{
\method{plot}{predictCox}(x, type = NULL, ci = FALSE, band = FALSE,
  groupBy = "row", reduce.data = FALSE, plot = TRUE, digits = 2,
  alpha = 0.1, ...)
}
\arguments{
\item{x}{object obtained with the function \code{predictCox}.}

\item{type}{the type of predicted value. Choices are \itemize{
\item \code{"hazard"} the baseline hazard function when
argument \code{newdata} is not used and the hazard function
when argument \code{newdata} is used.  \item \code{"cumhazard"}
the cumulative baseline hazard function when argument
\code{newdata} is not used and the cumulative hazard function
when argument \code{newdata} is used.  \item \code{"survival"}
the survival baseline hazard function when argument
\code{newdata} is not used and the cumulative hazard function
when argument \code{newdata} is used.  } Several choices can be
combined in a vector of strings that match (no matter the case)
strings \code{"hazard"},\code{"cumhazard"}, \code{"survival"}.}

\item{ci}{Logical. If \code{TRUE} display the confidence intervals for the predictions.}

\item{band}{Logical. If \code{TRUE} display the confidence bands for the predictions.}

\item{groupBy}{The grouping factor used to color the prediction curves. Can be \code{"row"}, \code{"strata"}, or \code{"covariates"}.}

\item{reduce.data}{Logical. If \code{TRUE} only the covariates that does take indentical values for all observations are displayed.}

\item{plot}{Logical. Should the graphic be plotted.}

\item{digits}{integer indicating the number of decimal places}

\item{alpha}{transparency of the confidence bands. Argument passed to \code{ggplot2::geom_ribbon}.}

\item{...}{not used. Only for compatibility with the plot method.}
}
\description{
Plot predictions from a Cox model
}
\examples{
library(survival)
library(ggplot2)

## predictions ##
d <- sampleData(1e2, outcome = "survival")
m.cox <- coxph(Surv(time,event)~ X1 + X2 + X3,
                data = d, x = TRUE, y = TRUE)
dt.basehaz <- predictCox(m.cox)
ggplot(as.data.table(dt.basehaz), aes(x = time, y = survival)) + geom_point() + geom_line()

pred.cox <- predictCox(m.cox, newdata = d[1:4,],
  times = 1:5, type = "survival", keep.newdata = TRUE)
plot(pred.cox)
plot(pred.cox, groupBy = "covariates")
plot(pred.cox, groupBy = "covariates", reduce.data = TRUE)


m.cox.strata <- coxph(Surv(time,event)~ strata(X1) + strata(X2) + X3 + X6,
data = d, x = TRUE, y = TRUE)
pred.cox.strata <- predictCox(m.cox.strata, newdata = d[c(1:5,10,50),],
time = 1:5, keep.newdata = TRUE)
plot(pred.cox.strata, type = "survival")
plot(pred.cox.strata, type = "survival", groupBy = "strata")
res <- plot(pred.cox.strata, type = "survival",
            groupBy = "covariates")

# customize display
res$plot + geom_point(data = res$dataDuplicated, size = 3)

## predictions with confidence interval
pred.cox <- predictCox(m.cox, newdata = d[1:4,],
  times = 1:5, type = "survival", se = TRUE, keep.newdata = TRUE)
plot(pred.cox, ci = TRUE)

## predictions with confidence bands
pred.cox <- predictCox(m.cox, newdata = d[1,,drop=FALSE],
  times = 1:5, type = "survival", nSim.band = 500, keep.newdata = TRUE)
plot(pred.cox, band = TRUE)


}

