% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{plot.score.AUC}
\alias{plot.score.AUC}
\title{Plot AUC curve}
\usage{
\method{plot}{score.AUC}(x, models, lwd = 3, xlim, ylim, axes = TRUE, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{models}{Choice of models to plot}

\item{lwd}{Line width}

\item{xlim}{Limits for x-axis}

\item{ylim}{Limits for y-axis}

\item{axes}{Logical. If \code{TRUE} draw axes.}

\item{...}{Not yet used}
}
\description{
Plot AUC curve
}

