% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{plot.score.Brier}
\alias{plot.score.Brier}
\title{Plot Brier curve}
\usage{
\method{plot}{score.Brier}(x, models, lwd = 3, xlim, ylim, axes = TRUE, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{models}{Choice of models to plot}

\item{lwd}{Line width}

\item{xlim}{Limits for x-axis}

\item{ylim}{Limits for y-axis}

\item{axes}{Logical. If \code{TRUE} draw axes.}

\item{...}{Not yet used}
}
\description{
Plot Brier curve
}

