% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scoreAUC.R
\name{plot.scoreAUC}
\alias{plot.scoreAUC}
\title{Plot AUC curve}
\usage{
\method{plot}{scoreAUC}(x, models, type = "score", lwd = 2, xlim, ylim, col,
  axes = TRUE, confint = FALSE, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{models}{Choice of models to plot}

\item{type}{Character. Either \code{"score"} to show AUC or \code{"test"} to show differences between AUC.}

\item{lwd}{Line width}

\item{xlim}{Limits for x-axis}

\item{ylim}{Limits for y-axis}

\item{col}{Color}

\item{axes}{Logical. If \code{TRUE} draw axes.}

\item{confint}{Logical. If \code{TRUE} draw confidence shadows.}

\item{...}{Not yet used}
}
\description{
Plot AUC curve
}

