% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scoreROC.R
\name{plot.scoreROC}
\alias{plot.scoreROC}
\title{Plot ROC curve}
\usage{
\method{plot}{scoreROC}(x, models, timepoint, lwd = 3, legend = TRUE,
  legend.title, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{models}{Choice of models to plot}

\item{timepoint}{Time point specifying the prediction horizon}

\item{lwd}{line width}

\item{legend}{Logical. If \code{TRUE} draw legend.}

\item{legend.title}{Legend title}

\item{...}{Not yet used}
}
\description{
Plot ROC curve
}

