% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBrier.R
\name{plotBrier}
\alias{plotBrier}
\title{Plot Brier curve}
\usage{
plotBrier(x, models, lwd = 3, xlim, ylim, axes = TRUE, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{models}{Choice of models to plot}

\item{lwd}{Line width}

\item{xlim}{Limits for x-axis}

\item{ylim}{Limits for y-axis}

\item{axes}{Logical. If \code{TRUE} draw axes.}

\item{...}{Not yet used}
}
\description{
Plot Brier score curves
}
\details{
Plot Brier score curves
}
\examples{
library(survival)
ds1=sampleData(40,outcome="survival")
ds2=sampleData(40,outcome="survival")
f1 <- coxph(Surv(time,event)~X1+X3+X5+X7+X9,data=ds1,x=TRUE)
f2 <- coxph(Surv(time,event)~X2+X4+X6+X8+X10,data=ds1,x=TRUE)
xscore <- Score(list(f1,f2),formula=Hist(time,event)~1,data=ds2,metrics="brier")
}

