% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRiskCI.R
\name{predictRiskCI}
\alias{predictRiskCI}
\alias{predictRiskCI.CauseSpecificCox}
\alias{predictRiskCI.coxph}
\alias{predictRiskCI.cph}
\title{Extrating predicting risks with confidence intervals from regression models}
\usage{
\method{predictRiskCI}{cph}(object,newdata,conf.level=0.95,times,colnames=TRUE,...)
\method{predictRiskCI}{coxph}(object,newdata,conf.level=0.95,times,colnames=TRUE,...)
\method{predictRiskCI}{CauseSpecificCox}(object,newdata,conf.level=0.95,times,cause,colnames=TRUE,...)
}
\arguments{
\item{object}{A fitted model from which to extract predicted event
probabilities}

\item{newdata}{A data frame containing predictor variable combinations for
which to compute predicted event probabilities.}

\item{conf.level}{confidence level of the interval.}

\item{...}{Additional arguments that are passed on to the current method.}

\item{times}{A vector of times in the range of the response variable, for
which the cumulative incidences event probabilities are computed.}

\item{cause}{Identifies the cause of interest among the competing events.}

\item{colnames}{Logical. If \code{TRUE} set the column names to he evaluation times in the output.}
}
\value{
A list containing the punctual estimate, the lower bound and the upper bound of the confidence interval.
Each element of the list is organised in the same way as the output of \code{predictRisk}.
}
\description{
Extract event probabilities with confidence intervals from fitted regression models and machine learning objects. 

The function predictRiskCI is a generic function, meaning that it invokes
specifically designed functions depending on the 'class' of the first
argument.

See \code{\link{predictRiskCI}}.
}
\examples{
set.seed(100)

#### survival ####
d <- sampleData(200,outcome="survival")
# fit a Cox model
library(rms)
cphmodel <- cph(Surv(time,event)~X1+X2,data=d,surv=TRUE,x=TRUE,y=TRUE)
library(survival)
coxphmodel <- coxph(Surv(time,event)~X1+X2,data=d,x=TRUE,y=TRUE)

# prediction with confidence interval
predictRiskCI(cphmodel, newdata = d[1:5,], times = 1:3)
predictRiskCI(coxphmodel, newdata = d[1:5,], times = 1:3)

#### competing risks ####
d <- sampleData(200,outcome="competing.risks")
CSCmodel <- CSC(Hist(time,event)~X1+X2, data = d)

predictRiskCI(CSCmodel, newdata = d[1:5,], times = 1:3, cause = 1)
}
\author{
Brice Ozenne broz@sund.ku.dk
}
\seealso{
See \code{\link{predictRisk}}.
}
\keyword{survival}

