% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{predictSurv}
\alias{predictSurv}
\title{Compute Event-Free Survival From a CSC Object}
\usage{
predictSurv(object, newdata, times, product.limit)
}
\arguments{
\item{object}{The fitted CSC object.}

\item{newdata}{[data.frame or data.table]  Contain the values of the predictor variables
defining subject specific predictions.
Should have the same structure as the data set used to fit the \code{object}.}

\item{times}{[numeric vector] Time points at which to return
the estimated survival.}

\item{product.limit}{[logical] If \code{TRUE} the survival is computed using the product limit estimator.}

\item{...}{not used.}
}
\description{
Compute event-free survival from a CSC object.
}
