% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ate.R
\name{print.ate}
\alias{print.ate}
\title{Print Average Treatment Effects}
\usage{
\method{print}{ate}(x, ...)
}
\arguments{
\item{x}{object obtained with function \code{ate}}

\item{...}{passed to summary}
}
\description{
Print average treatment effects.
}
\seealso{
\code{\link{summary.ate}} to obtained a more detailed output
\code{\link{confint.ate}} to compute confidence intervals/bands.
\code{\link{ate}} to compute the average treatment effects.
}
