% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ateCox.R
\name{print.ateCox}
\alias{print.ateCox}
\title{Print average treatment effects}
\usage{
\method{print}{ateCox}(x, digits = 3, ...)
}
\arguments{
\item{x}{object obtained with function \code{ateCSC}}

\item{digits}{Number of digits}

\item{...}{passed to print}
}
\description{
Print average treatment effects
}

