% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ateRobust.R
\name{print.ateRobust}
\alias{print.ateRobust}
\title{Print Average Treatment Effect}
\usage{
\method{print}{ateRobust}(x, digits = 3, efficient = TRUE,
  iid.full = x$iid, ...)
}
\arguments{
\item{x}{object obtained with the function \code{ateRobust}.}

\item{digits}{[integer, >0] indicating the number of decimal places.}

\item{efficient}{[logical] should the standard errors be corrected to improve efficiency
in presence of censoring.}

\item{iid.full}{[logical] should the standard errors
accounting for the uncertainty in the outcome and propensity score models be output?}

\item{...}{Passed to print.}
}
\description{
Print average treatment effect.
}
