% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.CauseSpecificCox.R
\name{print.CauseSpecificCox}
\alias{print.CauseSpecificCox}
\title{Print of a Cause-Specific Cox regression model}
\usage{
\method{print}{CauseSpecificCox}(x, ...)
}
\arguments{
\item{x}{Object obtained with CSC}

\item{...}{Passed to print}
}
\description{
Print of a Cause-Specific Cox regression model
}

