% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.FGR.R
\name{print.FGR}
\alias{print.FGR}
\title{Print of a Fine-Gray regression model}
\usage{
\method{print}{FGR}(x, ...)
}
\arguments{
\item{x}{Object fitted with function FGR}

\item{...}{passed to cmprsk::summary.crr}
}
\description{
Print of a Fine-Gray regression model
}
