% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.GLMnet.R
\name{print.GLMnet}
\alias{print.GLMnet}
\title{Print of a glmnet regression model}
\usage{
\method{print}{GLMnet}(x, ...)
}
\arguments{
\item{x}{Object obtained with GLMnet}

\item{...}{Passed to print}
}
\description{
Print of a penalized regression model which was fitted with \link{GLMnet}.
}
