% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Score.R
\name{print.highscore}
\alias{print.highscore}
\title{Print metric specific results of risk prediction assessment}
\usage{
\method{print}{highscore}(x, B = 0, digits = 3, ...)
}
\arguments{
\item{x}{Element of the result of \code{Score.list}}

\item{B}{Number of splits if any.}

\item{digits}{Number of digits}

\item{...}{passed to print}
}
\description{
Print metric specific element of risk prediction assessment
}

