% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.InfluenceCoxTest.R
\name{print.influenceCoxTest}
\alias{print.influenceCoxTest}
\title{Print the results of the influence test}
\usage{
\method{print}{influenceCoxTest}(x, ...)
}
\arguments{
\item{x}{Object obtained with influenceCoxTest}

\item{...}{Passed to print}
}
\description{
Print the results of the influence test
}
