% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.influenceTest.R
\name{print.influenceTest}
\alias{print.influenceTest}
\title{Output of the DIfference Between Two Estimates}
\usage{
\method{print}{influenceTest}(x, digits = 3, ...)
}
\arguments{
\item{x}{object obtained with the function \code{influenceTest}.}

\item{digits}{[integer, >0] indicating the number of decimal places.}

\item{...}{Passed to print.}
}
\description{
Output of the difference between two estimates.
}
\details{
to display confidence intervals/bands,
the \code{confint} method needs to be applied on the object.
}
\seealso{
\code{\link{confint.influenceTest}} to compute confidence intervals/bands.
\code{\link{influenceTest}} to perform the comparison.
}
