% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.predictCSC.R
\name{print.predictCSC}
\alias{print.predictCSC}
\title{Print Predictions From a Cause-specific Cox Proportional Hazard Regression}
\usage{
\method{print}{predictCSC}(x, digits = 3, ...)
}
\arguments{
\item{x}{object obtained with the function \code{predictCox}.}

\item{digits}{[integer, >0] indicating the number of decimal places.}

\item{...}{Passed to print.}
}
\description{
Print predictions from a Cause-specific Cox proportional hazard regression.
}
\details{
to display confidence intervals/bands,
the \code{confint} method needs to be applied on the object.
}
\seealso{
\code{\link{confint.predictCSC}} to compute confidence intervals/bands.
\code{\link{predict.CauseSpecificCox}} to compute the predicted risks.
}
