% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.riskRegression.R
\name{print.riskRegression}
\alias{print.riskRegression}
\title{Print function for riskRegression models}
\usage{
\method{print}{riskRegression}(x, times, digits = 3, eps = 10^-4,
  verbose = TRUE, ...)
}
\arguments{
\item{x}{Object obtained with ARR, LRR or riskRegression}

\item{times}{Time points at which to show time-dependent coefficients}

\item{digits}{Number of digits for all numbers but p-values}

\item{eps}{p-values smaller than this number are shown as such}

\item{verbose}{Level of verbosity}

\item{...}{not used}

\item{pvalue.digits}{Number of digits for p-values}
}
\description{
Print function for riskRegression models
}

