% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.subjectWeights.R
\name{print.subjectWeights}
\alias{print.subjectWeights}
\title{Print subject weights}
\usage{
\method{print}{subjectWeights}(x, digits = 3, ...)
}
\arguments{
\item{x}{Subject weights}

\item{digits}{Digits}

\item{...}{not used}
}
\description{
Print subject weights
}

