% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.synth_code.R
\name{print.synth_code}
\alias{print.synth_code}
\title{Print synthesized code}
\usage{
\method{print}{synth_code}(x, digits, ...)
}
\arguments{
\item{x}{Object obtained with \code{synthesize}}

\item{digits}{Number of digits. Not used.}

\item{...}{Not used.

method print synth_code}
}
\description{
Print method for synthesized code
}
