% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.timevarCoefList.R
\name{print.timevarCoefList}
\alias{print.timevarCoefList}
\title{Print time varying coefficients}
\usage{
\method{print}{timevarCoefList}(x, digits = 3, eps = 10^-4, ...)
}
\arguments{
\item{x}{object with time varying coefficients}

\item{digits}{Number of digits for all numbers but p-values}

\item{eps}{p-values smaller than this number are shown as such}

\item{...}{not used}
}
\description{
Print time varying coefficients
}

