% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskLevelPlot.R
\name{riskLevelPlot}
\alias{riskLevelPlot}
\title{Level plots for risk prediction models}
\usage{
riskLevelPlot(object, formula, data = parent.frame(), horizon = NULL,
  cause = 1, ...)
}
\arguments{
\item{object}{risk prediction model object}

\item{formula}{formula}

\item{data}{data}

\item{horizon}{time point}

\item{cause}{cause of interst}

\item{...}{passed to lattice::levelplot}
}
\description{
Level plots for predicted risks
}
\details{
Level plots for predicted risks
}
\examples{

# ---------- logistic regression --------------------
library(ModelGood)
expit <- function(x){exp(x)/(1+exp(x))}
partyData <- function(N){
  Age <- runif(N,.5,15)
  Parasites <- rnorm(N,mean=3.5-0.03*Age)
  Fever <- factor(rbinom(N,1,expit(-3.5-.3*Age+.55*Parasites+0.15*Age*Parasites)))
  data.frame(Fever,Age,Parasites)
}
d <- partyData(100)
f <- glm(Fever~Age+Parasites,data=d,family="binomial")
riskLevelPlot(f,Fever~Age+Parasites,d)
library(randomForest)
rf <- randomForest(Fever~Age+Parasites,data=d)
riskLevelPlot(f,Fever~Age+Parasites,d)
riskLevelPlot(rf,Fever~Age+Parasites,d)

# ---------- survival analysis --------------------

# --simulate an artificial data frame
# with survival response and three predictors

library(survival)
library(prodlim)
survData <- function( N ){
  sdata <- SimSurv( N )
  sdata$X3 <- rnorm(N)
  sdata
}
set.seed(140515)
sdat <- survData(100)
sdat <- sampleData(100,outcome="survival")
# -- fit a Cox regression model 
survForm = Surv(time,event) ~ X7 + X8
cox <- coxph(survForm, data = sdat,x=TRUE)

# --choose a time horizon for the predictions and plot the risks
timeHorizon <- floor(median(sdat$time))
riskLevelPlot(cox, survForm, data = sdat, horizon = timeHorizon)

survForm2 = Surv(time,status) ~ strata(X1) + X2 + X3
cox2 <- coxph(survForm2, data = sdat)
riskLevelPlot(cox2, survForm2, data = sdat, horizon = timeHorizon)

# ---------- competing risks --------------------

# -- simulate an artificial data frame
# with competing cause response and three predictors
library(cmprsk)
library(riskRegression)
compRiskData <- function( N ){
  crdata <- SimCompRisk( N )
  crdata$X3 <- rnorm(N)
  crdata
}
set.seed(140515)
crdat <- compRiskData(1000)

# -- fit a cause-specific Cox regression model
crForm <- Hist(time,event)~X2+X3
csCox  <- CSC(crForm, data=crdat)

# -- choose a time horizon and plot the risk for a given cause
timeHorizon <- floor(median(crdat$time))
riskLevelPlot(csCox, crForm, data = crdat, horizon = timeHorizon, cause = 1)

}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
