% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskRegression-package.R
\docType{package}
\name{riskRegression-package}
\alias{riskRegression-package}
\title{Risk Regression Models and Prediction Scores for Survival Analysis with Competing Risks}
\description{
Methods for evaluating risk predictions in censored event history analysis
}
\keyword{internal}
