% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0onload.R
\name{riskRegression.options}
\alias{riskRegression.options}
\title{Global options for \code{riskRegression}}
\usage{
riskRegression.options(...)
}
\arguments{
\item{...}{for now limited to \code{method.predictRisk} and \code{mehtod.predictRiskIID}.}
}
\description{
Output and set global options for the \code{riskRegression} package.
}
\details{
only used by the \code{ate} function.
}
\examples{
options <- riskRegression.options()

## add new method.predictRiskIID
riskRegression.options(method.predictRiskIID = c(options$method.predictRiskIID,"xx"))

riskRegression.options()
}
