% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowCumProd}
\alias{rowCumProd}
\title{Apply cumprod in each row}
\usage{
rowCumProd(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of t(apply(x,1,cumprod))
}
\examples{
x <- matrix(1:8,ncol=2)
rowCumProd(x)
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}
