% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rowCumSum}
\alias{rowCumSum}
\title{Apply cumsum in each row}
\usage{
rowCumSum(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix of same size as x.
}
\description{
Fast computation of t(apply(x,1,cumsum))
}
\examples{
x <- matrix(1:8,ncol=2)
rowCumSum(x)
}
\author{
Thomas Alexander Gerds <tag@biostat.ku.dk>
}
