% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rowPaste}
\alias{rowPaste}
\title{Collapse Rows of Characters.}
\usage{
rowPaste(object)
}
\arguments{
\item{object}{A matrix/data.frame/list containing the characters.}
}
\description{
Collapse rows of characters. Fast alternative to apply(x,1,paste0,collapse="")
}
\examples{
\dontrun{
M <- matrix(letters, nrow = 26, ncol = 2)
rowPaste(M)
}
}
