% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsquared.R
\name{rsquared}
\alias{rsquared}
\title{Explained variation for settings with binary, survival and competing risk outcome}
\usage{
rsquared(object, ...)
}
\arguments{
\item{object}{Model for which we want R^2}

\item{...}{passed to \code{riskRegression::Score}}
}
\value{
Data frame with explained variation values for the full model.
}
\description{
General R^2 for binary outcome and right censored time to event (survival) outcome also with competing risks
}
\details{
R^2 is calculated based on the model's predicted risks. The Brier score of the model is compared to the Brier score of the null model.
}
\examples{

# binary outcome
library(lava)
set.seed(18)
learndat <- sampleData(48,outcome="binary")
lr1 = glm(Y~X1+X2+X7+X9,data=learndat,family=binomial)
rsquared(lr1)

library(survival)
data(pbc)
cox1= coxph(Surv(time,status!=0)~age+sex+log(bili)+log(albumin)+log(protime),data=na.omit(pbc),x=TRUE)
rsquared(cox1,times=1000)

}
\seealso{
Score
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
