% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveCoxConfidential.R
\name{saveCoxConfidential}
\alias{saveCoxConfidential}
\title{library(survival)
library(lava)
set.seed(18)
trainSurv <- sampleData(300,outcome="survival")
testSurv <- sampleData(40,outcome="survival")
fit = coxph(Surv(time,event)~X1+X2+X3+X7+X9,data=trainSurv, y=TRUE, x = TRUE)
u=saveCoxConfidential(fit,times=3)
predictRisk(U,newdata=testsurv)
cox1 = coxph(Surv(time,event)~strata(X1)+X2+X3+X7+X9,data=trainSurv, y=TRUE, x = TRUE)
z<-saveCoxConfidential(cox1,c(2,5))}
\usage{
saveCoxConfidential(object, times)
}
\description{
dput(z) ## get output to copy object
}
\details{
all.equal(predictRisk(z,newdata=testSurv),
          predictRisk(cox1,newdata=testSurv,times=c(2,5)))

cox2 = coxph(Surv(time,event)~X1+X2+X7+X9,data=trainSurv, y=TRUE, x = TRUE)
z<-saveCoxConfidential(cox2,c(2,5))

predictRisk(z,testSurv)-predictRisk(z,testSurv,c(2,5))
}
