% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveSynth.R
\name{saveSynth}
\alias{saveSynth}
\title{Export a \code{synth} object.}
\usage{
saveSynth(object, file = "")
}
\arguments{
\item{object}{An object of type \code{synth}.}

\item{file}{File to print to.}
}
\value{
Either a string or nothing if printed to a file.
}
\description{
Function for exporting objects of type \code{synth} via \code{dput}.
}
\examples{
# See the documentation for the \code{synthesize} function.

}
\seealso{
lvm
}
