% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{score.wglm}
\alias{score.wglm}
\title{Score for IPCW Logistic Regressions}
\usage{
\method{score}{wglm}(x, indiv = FALSE, times = NULL, simplify = TRUE, ...)
}
\arguments{
\item{x}{a wglm object.}

\item{indiv}{[logical] should the individual score be output? Otherwise the total score (i.e. summed over all individuals will be output).}

\item{times}{[numeric vector] time points at which the score should be output.}

\item{simplify}{[logical] should the ouput be converted to a matrix when only one timepoint is requested. Otherwise will always return a list.}

\item{...}{Not used.}
}
\description{
Compute the first derivative of the log-likelihood for IPCW logistic regressions.
}
