% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{seCox}
\alias{seCox}
\title{Computation of standard errors for predictions}
\usage{
seCox(object, newdata, times, type, Lambda0, eXb, survival, stratavars)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{newdata}{A data frame or table containing the values of the
predictor variables defining subject specific predictions. Should have
the same structure as the data set used to fit the \code{object}.}

\item{times}{Time points at which to evaluate the predictions.}

\item{type}{One or several strings that match (either in lower or upper case or mixtures) one
or several of the strings \code{"hazard"},\code{"cumhazard"}, \code{"survival"}.}

\item{Lambda0}{the baseline hazard estimate returned by BaseHazStrata_cpp.}

\item{eXb}{exponential (covariates * coefficients)}

\item{survival}{the predicted survival (only necessary when type contains survival)}

\item{stratavars}{the name of the strata variables}
}
\value{
A list optionally containing the standard error for the survival, cumulative hazard and hazard.
}
\description{
Compute the standard error associated to the predictions from Cox regression model using the functional delta method.
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

