% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simActiveSurveillance.R
\name{simActiveSurveillance}
\alias{simActiveSurveillance}
\title{Simulate data of a hypothetical active surveillance prostate cancer study}
\usage{
simActiveSurveillance(n)
}
\arguments{
\item{n}{sample size}
}
\value{
data table of size n
}
\description{
Simulate data of a hypothetical active surveillance prostate cancer study
}
\details{
This is based on the functionality of \code{library(lava)}.
}
\examples{
set.seed(71)
simActiveSurveillance(3)
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
