% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMelanoma.R
\name{simMelanoma}
\alias{simMelanoma}
\title{Simulate data alike the Melanoma data}
\usage{
simMelanoma(n)
}
\arguments{
\item{n}{sample size}
}
\value{
data table of size n
}
\description{
Simulate data alike the Melanoma data
}
\details{
This is based on the functionality of \code{library(lava)}.
}
\examples{
set.seed(71)
simMelanoma(3)
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}

