% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.synth.R
\name{simsynth}
\alias{simsynth}
\title{Simulating from a synthesized object}
\usage{
simsynth(object, n = 200, drop.latent = FALSE, ...)
}
\arguments{
\item{object}{generated with \code{synthesize}}

\item{n}{sample size}

\item{drop.latent}{if \code{TRUE} remove the latent event times from the resulting data set.}

\item{...}{additional arguments passed on to \code{lava::sim}}
}
\description{
Simulating from a synthesized object
}
\examples{
library(survival)
m=synthesize(Surv(time,status)~sex+age+bili,data=pbc)
simsynth(m,10,drop.latent=TRUE)

}
