% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRisk.R
\name{SmcFcs}
\alias{SmcFcs}
\title{TODO}
\usage{
SmcFcs(formula, data, m = 5, method, fitter = "glm", fit.formula, ...)
}
\arguments{
\item{formula}{TODO}

\item{data}{TODO}

\item{m}{TODO}

\item{method}{TODO}

\item{fitter}{TODO}

\item{fit.formula}{TODO}

\item{...}{TODO}
}
\description{
TODO
}
