% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ate.R
\name{summary.ate}
\alias{summary.ate}
\title{Summary Average Treatment Effects}
\usage{
\method{summary}{ate}(
  object,
  estimator = object$estimator[1],
  type = c("meanRisk", "diffRisk", "ratioRisk"),
  se = object$se,
  quantile = FALSE,
  digits = 3,
  ...
)
}
\arguments{
\item{object}{object obtained with function \code{ate}}

\item{estimator}{[character] The type of estimator relative to which the estimates should be displayed.}

\item{type}{[character vector] what to displayed.
Can be any combination of \code{"meanRisk"}, \code{"diffRisk"}, and \code{"ratioRisk"}.}

\item{se}{[logical] should the standard error of the risks be displayed?}

\item{quantile}{[logical] should the quantile of the confidence bands be displayed?}

\item{digits}{[integer, >0] Number of digits.}

\item{...}{passed to print}
}
\description{
Summary average treatment effects.
}
\details{
to display confidence intervals/bands and p.value,
the \code{confint} method needs to be applied on the object.
}
\seealso{
\code{\link{confint.ate}} to compute confidence intervals/bands.
\code{\link{ate}} to compute the average treatment effects.
}
