% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.FGR.R
\name{summary.FGR}
\alias{summary.FGR}
\title{Summary of a Fine-Gray regression model}
\usage{
\method{summary}{FGR}(object, ...)
}
\arguments{
\item{object}{Object fitted with function FGR}

\item{...}{passed to cmprsk::summary.crr}
}
\description{
Summary of a Fine-Gray regression model
}
