% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.riskRegression.R
\name{summary.riskRegression}
\alias{summary.riskRegression}
\title{Summary of a risk regression model}
\usage{
\method{summary}{riskRegression}(
  object,
  times,
  digits = 3,
  pvalue.digits = 4,
  eps = 10^-4,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{object}{Object obtained with ARR, LRR or riskRegression}

\item{times}{Time points at which to show time-dependent coefficients}

\item{digits}{Number of digits for all numbers but p-values}

\item{pvalue.digits}{Number of digits for p-values}

\item{eps}{p-values smaller than this number are shown as such}

\item{verbose}{Level of verbosity}

\item{...}{not used}
}
\description{
Summary of a risk regression model
}
