% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.Score.R
\name{summary.Score}
\alias{summary.Score}
\title{Summary of prediction performance metrics}
\usage{
\method{summary}{Score}(
  object,
  times,
  what = c("score", "contrasts"),
  models,
  digits = 1,
  pvalue.digits = 4,
  ...
)
}
\arguments{
\item{object}{Object obtained with \code{Score}.}

\item{times}{Select time points}

\item{what}{Either \code{"score"}, \code{"contrasts"} or both, i.e., \code{c("score","contrasts")}}

\item{models}{Select which models to summarize. Need to be a subset of \code{object$models}}

\item{digits}{For rounding everything but p-values}

\item{pvalue.digits}{For rounding p-values}

\item{...}{not used}
}
\value{
List of tables
}
\description{
Summarizing a Score object
}
\details{
The AUC and the Brier score are put into tables
}
\seealso{
Score
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
