% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{SurvResponseVar}
\alias{SurvResponseVar}
\title{Extract the time and event variable from a Cox model}
\usage{
SurvResponseVar(formula)
}
\arguments{
\item{formula}{a formula}
}
\description{
Extract the time and event variable from a Cox model
}
\author{
Brice Ozenne broz@sund.ku.dk
}
