% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{terms.phreg}
\alias{terms.phreg}
\title{Extract terms for phreg objects}
\usage{
\method{terms}{phreg}(x, ...)
}
\arguments{
\item{x}{a phreg object.}

\item{...}{not used.}
}
\description{
Extract terms for phreg objects
}
