% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformT}
\alias{transformT}
\title{Compute T-statistic after a Transformation}
\usage{
transformT(estimate, se, null, type, alternative)
}
\arguments{
\item{estimate}{[numeric matrix] the estimate value before transformation.}

\item{se}{[numeric matrix] the standard error after transformation.}

\item{null}{[numeric] the value of the estimate (before transformation) under the null hypothesis.}

\item{type}{[character] the transforamtion.
Can be \code{"none"}, \code{"log"}, \code{"loglog"}, \code{"cloglog"}, or \code{"atanh"} (Fisher transform).}

\item{alternative}{[character] a character string specifying the alternative hypothesis,
must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}.}
}
\description{
Compute T-statistic after a transformation.
}
\details{
\code{se} and \code{estimate} must have same dimensions.
}
