% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{vcov.wglm}
\alias{vcov.wglm}
\title{Variance-covariance for IPCW Logistic Regressions}
\usage{
\method{vcov}{wglm}(object, times = NULL, simplify = TRUE, ...)
}
\arguments{
\item{times}{[numeric vector] time points at which the variance-covariance matrix should be output.}

\item{simplify}{[logical] should the ouput be converted to a matrix when only one timepoint is requested. Otherwise will always return a list.}

\item{...}{Not used.}

\item{x}{a wglm object.}
}
\description{
Compute the variance-covariance matrix of the estimated model parameters of IPCW logistic regressions.
}
