% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights.R
\name{weights.ate}
\alias{weights.ate}
\title{Extract Inverse Probability Weights}
\usage{
\method{weights}{ate}(object, type = "IPTW", ...)
}
\arguments{
\item{object}{A \code{ate} object, i.e. output of the \code{ate} function.}

\item{type}{[character] type of weight to be extracted: \itemize{
\item \code{"probaT"}: probability that a given subject (in rows) get a given treatment (in columns).
\item \code{"probaC"}: probability that a given subject (in rows) is still at risk at a given time (in columns).
\item \code{"IPTW"}: observed treatment indicator divided by probaT.
\item \code{"IPCW"}: observed at risk indicator divided by probaT.
\item \code{"IPTWbox"}: create an object containing the weights and corresponding influence function (if exported by ate) that can be passed to the argument \code{treatment} of the \code{ate} function.
}}

\item{...}{Not used. For compatibility with the generic method.}
}
\value{
A matrix with as many lines as observations and as many columns as treatment groups (\code{"probaT"} or \code{"IPTW"})
or as timepoint (\code{"probaC"} or \code{"IPCW"}) at which to evaluate average treatment effects.
}
\description{
Extract weights used to re-balance treatment groups (IPTW) and handle right-censoring (IPCW).
}
\details{
Relevant when using IPTW or AIPTW estimators but not for G-formula estimators as no weights is involved.
}
