% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.ate.R
\name{as.data.table.ate}
\alias{as.data.table.ate}
\title{Turn ate Object Into a \code{data.table}}
\usage{
\method{as.data.table}{ate}(
  x,
  keep.rownames = FALSE,
  estimator = x$estimator,
  type = c("meanRisk", "diffRisk", "ratioRisk"),
  ...
)
}
\arguments{
\item{x}{object obtained with function \code{ate}}

\item{keep.rownames}{Not used.}

\item{estimator}{[character] The type of estimator relative to which the estimates should be output.}

\item{type}{[character vector] The type of risk to export.
Can be \code{"meanRisk"} to export the risks specific to each treatment group,
\code{"diffRisk"} to export the difference in risks between treatment groups,
or \code{"ratioRisk"} to export the ratio of risks between treatment groups.}

\item{...}{Not used.}
}
\description{
Turn ate object into a \code{data.table}.
}
