% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.ate.R
\name{as.data.table.ate}
\alias{as.data.table.ate}
\title{Turn ate Object Into a \code{data.table}}
\usage{
\method{as.data.table}{ate}(
  x,
  estimator = x$estimator,
  se = x$se,
  ci = x$ci,
  band = x$band,
  quantileBand = x$band,
  p.value = x$p.value,
  keep.rownames = FALSE,
  ...
)
}
\arguments{
\item{x}{object obtained with function \code{ate}}

\item{estimator}{[character] The type of estimator relative to which the estimates should be output.}

\item{se}{[logical] Should the standard errors be output?}

\item{ci}{[logical] Should the confidence intervals be output?}

\item{band}{[logical] Should the confidence bands be output?}

\item{quantileBand}{[logical] Should the p-values/adjusted p-values be output?}

\item{keep.rownames}{Not used.}

\item{...}{Not used.}
}
\description{
Turn ate object into a \code{data.table}.
}
