% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.ate.R
\name{as.data.table.ate}
\alias{as.data.table.ate}
\title{Turn ate Object Into a \code{data.table}}
\usage{
\method{as.data.table}{ate}(
  x,
  keep.rownames = FALSE,
  se = TRUE,
  estimator = x$estimator,
  ...
)
}
\arguments{
\item{x}{object obtained with function \code{ate}}

\item{keep.rownames}{Not used.}

\item{se}{[logical] Should standard errors/quantile for confidence bands be displayed?}

\item{estimator}{[character] The type of estimator relative to which the estimates should be output.}

\item{...}{Not used.}
}
\description{
Turn ate object into a \code{data.table}.
}
