% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.ateRobust.R
\name{as.data.table.ateRobust}
\alias{as.data.table.ateRobust}
\title{Turn ate Object Into a \code{data.table}}
\usage{
\method{as.data.table}{ateRobust}(x, keep.rownames = FALSE, ...)
}
\arguments{
\item{x}{object obtained with function \code{ate}}

\item{keep.rownames}{Not used.}

\item{...}{Not used.}
}
\description{
Turn ate object into a \code{data.table}.
}
