% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.predictCox.R
\name{as.data.table.predictCox}
\alias{as.data.table.predictCox}
\title{Turn predictCox Object Into a \code{data.table}}
\usage{
\method{as.data.table}{predictCox}(x, keep.rownames = FALSE, se = TRUE,
  ...)
}
\arguments{
\item{x}{object obtained with function \code{predictCox}}

\item{keep.rownames}{Not used.}

\item{se}{[logical] Should standard errors/quantile for confidence bands be displayed?}

\item{...}{Not used.}
}
\description{
Turn predictCox object into a \code{data.table}.
}
