% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.table.predictCSC.R
\name{as.data.table.predictCSC}
\alias{as.data.table.predictCSC}
\title{Turn predictCSC Object Into a \code{data.table}}
\usage{
\method{as.data.table}{predictCSC}(x, keep.rownames = FALSE, se = TRUE,
  ...)
}
\arguments{
\item{x}{object obtained with function \code{predictCSC}}

\item{keep.rownames}{not used}

\item{se}{should standard errors/quantile for confidence bands be displayed?}

\item{...}{not used}
}
\description{
Turn predictCSC object into a \code{data.table}.
}
