% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateCSC.R
\name{ateCSC}
\alias{ateCSC}
\title{Compute the average treatment effect using CSC

Use the g-formula to estimate the average treatment effect}
\usage{
ateCSC(data, treatVar, treatLevels = NULL, ls.formula, times, cause,
  return.model = FALSE, addTime = TRUE, n.bootstrap = 0, ncpus = 1,
  seed = NULL, alpha = 0.05, ...)
}
\arguments{
\item{data}{dataset}

\item{treatVar}{name of the column containing the treatment variable}

\item{treatLevels}{the levels of treatment to be compared}

\item{ls.formula}{formula argument of CSC}

\item{times}{time argument of CSC}

\item{cause}{argument of CSC}

\item{return.model}{should the fitted CSC model be returned}

\item{addTime}{should the time be specified in the output}

\item{n.bootstrap}{the number of bootstrap replications used to compute the confidence interval}

\item{ncpus}{the number of CPU to use for the bootstrap}

\item{seed}{the seed to set at the beginning of the bootstrap}

\item{alpha}{the type 1 error}

\item{...}{passed on}
}
\value{
A list with:
model: the CSC model
averageCIF: the predicted CIF averaged over patient for a given treatment modality (column) and time (row)
gFormula: A list (times) containing the value of the g formula comparing all pair of treatment modality, with the first one as a reference.
averageCIFCIsup: the lower bound of the CI for averageCIF
averageCIFCIinf: the upper bound of the CI for averageCIF
gFormulaCIsup: the lower bound of the CI for averageCIFCIsup
gFormulaCIinf: the upper bound of the CI for averageCIFCIinf
}
\details{
Require the snowfall package for the bootstrap.
}
\examples{
#library(snowfall)
data <- SimCompRisk(1e3)
data$time <- round(data$time,1)

seqtimes <- sample(x = unique(sort(data$time)), size = 100) 

data$X1 <- factor(rbinom(1e3, prob = c(0.2,0.3,0.2) , size = 3), labels = paste0("T",0:3))
res.Gformula <- ateCSC(data = data, treatVar = "X1", treatLevels = NULL,
                         ls.formula = Hist(time,event)~ X1*X2, 
                         times = 1:10, cause = 1,
                         n.bootstrap = 10)
 paste0(res.Gformula$averageCIF[res.Gformula$averageCIF[,"time"] == 5,"T0"],
       " [",
        res.Gformula$averageCIFCIinf[res.Gformula$averageCIFCIinf[,"time"] == 5,"T0"],
        " ; ",
        res.Gformula$averageCIFCIsup[res.Gformula$averageCIFCIsup[,"time"] == 5,"T0"],
        "]")
}

