% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.ate.R
\name{autoplot.ate}
\alias{autoplot.ate}
\title{Plot Average Risks}
\usage{
\method{autoplot}{ate}(
  object,
  estimator = object$estimator[1],
  ci = object$ci,
  band = object$band,
  plot = TRUE,
  smooth = FALSE,
  digits = 2,
  alpha = NA,
  ...
)
}
\arguments{
\item{object}{Object obtained with the function \code{ate}.}

\item{estimator}{[character] The type of estimator relative to which the risks should be displayed.}

\item{ci}{[logical] If \code{TRUE} display the confidence intervals for the average risks.}

\item{band}{[logical] If \code{TRUE} display the confidence bands for the average risks.}

\item{plot}{[logical] Should the graphic be plotted.}

\item{smooth}{[logical] Should a smooth version of the risk function be plotted instead of a simple function?}

\item{digits}{[integer, >0] Number of decimal places.}

\item{alpha}{[numeric, 0-1] Transparency of the confidence bands. Argument passed to \code{ggplot2::geom_ribbon}.}

\item{...}{not used. Only for compatibility with the plot method.}
}
\value{
Invisible. A list containing:
\itemize{
\item plot: the ggplot object.
\item data: the data used to create the plot.
}
}
\description{
Plot average risks.
}
\examples{
library(survival)
library(rms)
library(ggplot2)

#### simulate data ####
n <- 1e2
set.seed(10)
dtS <- sampleData(n,outcome="survival")
seqTimes <- c(0,sort(dtS$time[dtS$event==1]),max(dtS$time))

#### Cox model ####
fit <- cph(formula = Surv(time,event)~ X1+X2,data=dtS,y=TRUE,x=TRUE)

#### Average treatment effect ####
ateFit <- ate(fit, data = dtS, treatment = "X1",
              times = seqTimes[seqTimes>5], se = TRUE, band = TRUE)

#### display #### 
ggplot2::autoplot(ateFit)
outGG <- ggplot2::autoplot(ateFit, alpha = 0.1)

dd <- as.data.frame(outGG$data[treatment == 0])
outGG$plot + facet_wrap(~treatment, labeller = label_both)
}
\seealso{
\code{\link{ate}} to compute average risks.
}
