% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Score.R
\name{boxplot.Score}
\alias{boxplot.Score}
\title{Boxplot risk quantiles}
\usage{
\method{boxplot}{Score}(x, model, reference, type, timepoint, lwd = 3, xlim,
  xlab, main, ...)
}
\arguments{
\item{x}{Score object obtained by calling function \code{Score}.}

\item{model}{Choice of risk prediction model}

\item{reference}{Choice of reference risk prediction model for
calculation of risk differences.}

\item{type}{Either \code{"risk"} for predicted risks or
\code{"diff"} for differences between predicted risks.}

\item{timepoint}{time point specifying the prediction horizon}

\item{lwd}{line width}

\item{xlim}{x-axis limits}

\item{xlab}{x-axis label}

\item{main}{title of plot}

\item{...}{not used}
}
\description{
Retrospective boxplots of risk quantiles conditional on outcome
}
\examples{
db=sampleData(100,outcome="binary")
fitconv=glm(Y~X3+X5,data=db,family=binomial)
fitnew=glm(Y~X1+X3+X5+X6+X7,data=db,family=binomial)
scoreobj=Score(list(new=fitnew,conv=fitconv),formula=Y~1,
               data=db,summary="riskQuantile",nullModel=FALSE)
boxplot(scoreobj)

library(survival)
ds=sampleData(100,outcome="survival")
fitconv=coxph(Surv(time,event)~X3+X5,data=ds)
fitnew=coxph(Surv(time,event)~X1+X3+X5+X6+X7,data=ds)
scoreobj=Score(list(conv=fitconv,new=fitnew),formula=Hist(time,event)~1,
               data=ds,summary="riskQuantile",times=5,nullModel=FALSE)
boxplot(scoreobj)

library(riskRegression)
data(Melanoma)
fitconv = CSC(Hist(time,status)~invasion+age+sex+logthick,data=Melanoma)
fitnew = CSC(Hist(time,status)~invasion+age+sex,data=Melanoma)
scoreobj=Score(list(conv=fitconv,new=fitnew),formula=Hist(time,status)~1,
               data=Melanoma,summary="riskQuantile",times=5*365.25,nullModel=FALSE)
boxplot(scoreobj)
}

