% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.CauseSpecificCox.R
\name{coef.CauseSpecificCox}
\alias{coef.CauseSpecificCox}
\title{Extract coefficients from a Cause-Specific Cox regression model}
\usage{
\method{coef}{CauseSpecificCox}(x)
}
\arguments{
\item{x}{Object obtained with CSC}
}
\description{
Extract coefficients from a Cause-Specific Cox regression model
}

