% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{coef.wglm}
\alias{coef.wglm}
\title{Coefficient of IPCW Logistic Regressions}
\usage{
\method{coef}{wglm}(object, times = NULL, simplifies = TRUE, ...)
}
\arguments{
\item{times}{[numeric vector] time points at which the coefficients should be output.}

\item{simplifies}{[logical] should the ouput be converted to a vector when only one timepoint is requested. Otherwise will always return a matrix.}

\item{...}{Not used.}

\item{x}{a wglm object.}
}
\description{
Extract regression coefficients (relative to the outcome model) of a IPCW Logisitc model.
}
