% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{confint.wglm}
\alias{confint.wglm}
\title{Confidence intervals for Estimate from IPCW Logistic Regressions}
\usage{
\method{confint}{wglm}(object, parm = NULL, level = 0.95, times = NULL, type = "robust", ...)
}
\arguments{
\item{object}{a wglm object.}

\item{parm}{not used. For compatibility with the generic method.}

\item{level}{[numeric, 0-1] Level of confidence.}

\item{times}{[numeric vector] time points at which the estimates should be output.}

\item{type}{[character] should the robust variance-covariance matrix be computing accounting for the uncertainty of the IPCW (\code{"robust"})
or ignoring the uncertainty of the IPCW (\code{"robust-wknown"}), or model-based ignoring the uncertainty of the IPCW  (\code{"model-wknown"})?}

\item{...}{Not used.}
}
\description{
Display the confidence intervals w.r.t. the estimated regression parameters from IPCW logistic regressions.
}
