% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxCenter}
\alias{coxCenter}
\alias{coxCenter.cph}
\alias{coxCenter.coxph}
\alias{coxCenter.phreg}
\title{Extract the mean value of the covariates}
\usage{
coxCenter(object)

\method{coxCenter}{cph}(object)

\method{coxCenter}{coxph}(object)

\method{coxCenter}{phreg}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}
}
\description{
Extract the mean value of the covariates
}
\author{
Brice Ozenne broz@sund.ku.dk
}
