% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxData}
\alias{CoxData}
\alias{CoxData.coxph}
\alias{CoxData.cph}
\title{Extract the dataset used to train a Cox model}
\usage{
CoxData(object)

\method{CoxData}{cph}(object)

\method{CoxData}{coxph}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}
}
\description{
Extract the dataset used to train a Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
CoxData(mCox)

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
CoxData(mCox)
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

