% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxFormula}
\alias{coxFormula}
\alias{coxFormula.cph}
\alias{coxFormula.coxph}
\alias{coxFormula.phreg}
\title{Extract the formula from a Cox model}
\usage{
coxFormula(object)

\method{coxFormula}{cph}(object)

\method{coxFormula}{coxph}(object)

\method{coxFormula}{phreg}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}
}
\description{
Extract the formula from a Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mcox <- coxph(Surv(time, event) ~ X1+X2, data = d)
coxFormula(mcox)

##
library(rms)
f <- Surv(time, event) ~ X1+X2
mcox <- cph(f, data = d, y = TRUE)
coxFormula(mcox)
}
}
\author{
Brice Ozenne broz@sund.ku.dk
}
