% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxLP}
\alias{CoxLP}
\alias{CoxLP.coxph}
\alias{CoxLP.cph}
\title{Compute the linear predictor of a Cox model}
\usage{
CoxLP(object, data, lpVars, stratavars, center)

\method{CoxLP}{cph}(object, data = NULL, lpVars, stratavars, center)

\method{CoxLP}{coxph}(object, data = NULL, lpVars, stratavars, center)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{data}{a \code{data.frame} or a \code{data.table}}

\item{lpVars}{the name of all the covariates}

\item{stratavars}{the variables used to define the strata}

\item{center}{should the linear predictor be computed after centering the covariates}
}
\description{
Compute the linear predictor of a Cox model
}
\details{
In case of empty linear predictor returns a vector of 0 with the same length as the number of rows of the dataset
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
CoxLP(mCox, data = d, xvars = c("X1","X2"), stratavars = NULL, center = FALSE)
CoxLP(mCox, data = d, xvars = c("X1","X2"), stratavars = NULL, center = TRUE)  
mCoxS <- coxph(Surv(time, event) ~ strata(X1)+X2, data = d)
CoxLP(mCoxS, data = d, xvars = c("X1","X2"), stratavars = "strata(X1)", center = FALSE) 
mCoxS2 <- coxph(Surv(time, event) ~ strata(X1)+strata(X2), data = d)
CoxLP(mCoxS2, data = d, xvars = c("X1","X2"), stratavars = c("X1","X2"), center = FALSE) 

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
CoxLP(mCox, data = d, xvars = c("X1","X2"), stratavars = NULL, center = FALSE) 
CoxLP(mCox, data = d, xvars = c("X1","X2"), stratavars = NULL, center = TRUE) 
mCoxS <- cph(Surv(time, event) ~ strat(X1)+X2, data = d)
CoxLP(mCoxS, data = d, xvars = c("X1","X2"), stratavars = "strat(X1)", center = FALSE) 
mCoxS2 <- cph(Surv(time, event) ~ strat(X1)+strat(X2), data = d, y = TRUE)
CoxLP(mCoxS2, data = d, xvars = c("X1","X2"), stratavars = c("X1","X2"), center = FALSE) 
}
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

