% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{CoxN}
\alias{CoxN}
\alias{CoxN.coxph}
\alias{CoxN.cph}
\title{Extract the number of observations from a Cox model}
\usage{
CoxN(object)

\method{CoxN}{cph}(object)

\method{CoxN}{coxph}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}
}
\description{
Extract the number of observations from a Cox model
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")

##
library(survival)
mCox <- coxph(Surv(time, event) ~ X1+X2, data = d)
CoxN(mCox)

##
library(rms)
mCox <- cph(Surv(time, event) ~ X1+X2, data = d, y = TRUE)
CoxN(mCox)
}
}
\author{
Brice Ozenne broz@sund.ku.dk
}

