% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxN}
\alias{coxN}
\alias{coxN.cph}
\alias{coxN.coxph}
\alias{coxN.phreg}
\title{Extract the number of observations from a Cox model}
\usage{
coxN(object)

\method{coxN}{cph}(object)

\method{coxN}{coxph}(object)

\method{coxN}{phreg}(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}
}
\description{
Extract the number of observations from a Cox model
}
\author{
Brice Ozenne broz@sund.ku.dk
}
