% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxStrata}
\alias{coxStrata}
\alias{coxStrata.cph}
\alias{coxStrata.coxph}
\alias{coxStrata.phreg}
\title{Define the strata for a new dataset}
\usage{
coxStrata(object, data, sterms, strata.vars, levels, strata.levels)

\method{coxStrata}{coxph}(object, data, sterms, strata.vars, levels,
  strata.levels)

\method{coxStrata}{coxph}(object, data, sterms, strata.vars, levels,
  strata.levels)

\method{coxStrata}{phreg}(object, data, sterms, strata.vars, levels,
  strata.levels)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package), \code{cph}
(rms package), or \code{phreg} (mets package).}

\item{data}{a \code{data.frame} or a \code{data.table}}

\item{sterms}{terms in the formula corresponding to the strata variables}

\item{strata.vars}{the name of the variables used to define the strata}

\item{levels}{the strata levels that have been used to fit the Cox model}

\item{strata.levels}{a named list containing for each variable used to form the strata all its possible levels}
}
\description{
Define the strata in a dataset to match those of a stratified Cox model
}
\details{
if no strata variables returns a vector of \code{"1"} (factor).
}
\examples{
\dontrun{
d <- sampleData(1e2, outcome = "survival")
d$entry <- 0
d$id <- 1:NROW(d)

##
library(survival)
mcoxS <- coxph(Surv(time, event) ~ strata(X1)+strata(X2), data = d, x = TRUE, y = TRUE)
resInfo <- coxVariableName(mcoxS)
Ostrata <- coxStrata(mcoxS, strata.vars = resInfo$strata.vars)
coxStrata(mcoxS, data = d, sterms = resInfo$sterms, strata.vars = resInfo$strata.vars, 
          levels = levels(Ostrata), strata.levels = resInfo$strata.levels)

##
library(rms)
mcoxS <- cph(Surv(time, event) ~ strat(X1)+strat(X2), data = d, y = TRUE)
resInfo <- coxVariableName(mcoxS)
Ostrata <- coxStrata(mcoxS, strata.vars = resInfo$strata.vars)
coxStrata(mcoxS, data = d, sterms = resInfo$sterms, strata.vars = resInfo$strata.vars, 
          levels = levels(Ostrata), strata.levels = resInfo$strata.levels)
          
##
library(mets)
mcoxS <- phreg(Surv(entry, time, event) ~ strata(X1)+X2+cluster(id), data = d)
resInfo <- coxVariableName(mcoxS)
Ostrata <- coxStrata(mcoxS, strata.vars = resInfo$strata.vars)
coxStrata(mcoxS, data = d, sterms = resInfo$sterms, strata.vars = resInfo$strata.vars, 
          levels = levels(Ostrata), strata.levels = resInfo$strata.levels)
}
}
\author{
Brice Ozenne broz@sund.ku.dk
}
