% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{coxVariableName}
\alias{coxVariableName}
\title{Extract variable names from a model}
\usage{
coxVariableName(object)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}
}
\description{
Extract the name of the variables belonging to the linear predictor or used to form the strata
}
\author{
Brice Ozenne broz@sund.ku.dk
}
