% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLMnet.R, R/GMLnet.R
\name{GLMnet}
\alias{GLMnet}
\title{Fitting GLMnet for use with predictRisk}
\usage{
GLMnet(
  formula,
  data,
  lambda = NULL,
  cv = TRUE,
  alpha = 1,
  nfolds = 10,
  type.measure = "deviance",
  ...
)

GLMnet(
  formula,
  data,
  lambda = NULL,
  cv = TRUE,
  alpha = 1,
  nfolds = 10,
  type.measure = "deviance",
  ...
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{The data on which to fit the model.}

\item{lambda}{The tuning parameters for GLMnet. If set to NULL, then it the parameters are chosen for you.}

\item{cv}{Whether to use cross-validation or not. Default is TRUE.}

\item{alpha}{The elasticnet mixing parameter. See the ?glmnet for more details.}

\item{nfolds}{Number of folds for cross-validation. Default is 10.}

\item{type.measure}{loss to use for cross-validation. Default is deviance.}

\item{\dots}{Additional arguments that are passed on to the glmnet.}
}
\description{
Fit GLMnet models via a formula and a data set for use with \code{\link{predictRisk}}.

Fit GLMnet models via a formula and a data set for use with \code{\link{predictRisk}}.
}
