% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenceCox.R
\name{iid}
\alias{iid}
\alias{iid.coxph}
\title{Extract i.i.d. decomposition from a Cox model}
\usage{
iid(x, ...)

\method{iid}{coxph}(x, tau = NULL, ...)
}
\arguments{
\item{x}{object The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{...}{additional arguments}
}
\description{
Compute the influence function for each observation used to estimate the model
}
\examples{
library(survival)
library(data.table)
set.seed(10)
d <- sampleData(2e1, outcome = "survival")[,.(eventtime,event,X1,X6)]
setkey(d, eventtime)

library(timereg)
mGS.cox <- cox.aalen(Surv(eventtime, event) ~ prop(X1)+prop(X6), data = d, resample.iid = TRUE)
ICbeta_GS <- mGS.cox$gamma.iid
IClambda_GS <- t(as.data.table(mGS.cox$B.iid))
 
 
m.cox <- coxph(Surv(eventtime, event) ~ X1+X6, data = d, y = TRUE)
IC.cox <- iid.coxph(m.cox, center = FALSE)

range(IC.cox$ICbeta-ICbeta_GS)
range(IC.cox$ICLambda0/IClambda_GS[,NCOL(IClambda_GS)])

var(ICbeta_GS)
m.cox[["var"]]
var(ICbeta_GS)/mGS.cox$var.gamma


var(IC.cox$IC)/m.cox$var

var(mGS.cox$gamma.iid)
m.cox$var

var(mGS.cox$gamma.iid)/m.cox$var


 
IC.cox$epsilon11/mGS.cox$gamma.iid

}

