% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{iid.wglm}
\alias{iid.wglm}
\title{IID for IPCW Logistic Regressions}
\usage{
iid.wglm(x, times = NULL, simplifies = TRUE, ...)
}
\arguments{
\item{x}{a wglm object.}

\item{times}{[numeric vector] time points at which the iid should be output.}

\item{simplifies}{[logical] should the ouput be converted to a matrix when only one timepoint is requested. Otherwise will always return a list.}

\item{...}{Not used.}
}
\description{
Compute the decomposition in iid elements of the ML estimor of IPCW logistic regressions.
}
