% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influenceTest.R
\name{influenceTest}
\alias{influenceTest}
\alias{influenceCoxTest}
\alias{influenceCoxTest.list}
\alias{influenceCoxTest.default}
\title{Influence test [Experimental!!]}
\usage{
influenceCoxTest(object, ...)

\method{influenceCoxTest}{list}(object, newdata, times, type, cause, ...)

\method{influenceCoxTest}{default}(object, object2, nSim.band = 10000, ...)
}
\arguments{
\item{object}{either a list of models or an object of class predictCox or predictCSC.}

\item{...}{additional arguments to be passed to lower level functions.}

\item{newdata}{A \code{data.frame} or \code{data.table} containing
the values of the predictor variables defining subject specific
predictions.}

\item{times}{Time points at which to evaluate the predictions.}

\item{type}{the type of predicted value.}

\item{cause}{Identifies the cause of interest among the competing
events.}

\item{object2}{same as predict1 but for another model.}

\item{nSim.band}{the number of simulations used to compute the quantiles
for the confidence bands.}
}
\description{
Compare two estimates using their influence function
}
\examples{
library(lava)
library(survival)
n <- 100

#### Under H1
set.seed(1)
newdata <- data.frame(X1=1)

## simulate non proportional hazard using lava
m <- lvm()
regression(m) <- y ~ 1
regression(m) <- s ~ exp(-2*X1)
distribution(m,~X1) <- binomial.lvm()
distribution(m,~cens) <- coxWeibull.lvm(scale=1)
distribution(m,~y) <- coxWeibull.lvm(scale=1,shape=~s)
eventTime(m) <- eventtime ~ min(y=1,cens=0)
d <- as.data.table(sim(m,n))
setkey(d, eventtime)

## fit cox models
m.cox <- coxph(Surv(eventtime, status) ~ X1, 
               data = d, y = TRUE, x = TRUE)

mStrata.cox <- coxph(Surv(eventtime, status) ~ strata(X1), 
                     data = d, y = TRUE, x = TRUE)

## compare models
influenceCoxTest(list(m.cox, mStrata.cox),
                 type = "survival", newdata = newdata, times = seq(0.1,1,by=0.1))

#### Under H0 (Cox) ####
set.seed(1)
d <- sampleData(n, outcome = "survival")
setkey(d, eventtime)

## fit cox models
m.cox <- coxph(Surv(eventtime, event) ~ X1, 
               data = d, y = TRUE, x = TRUE)

mStrata.cox <- coxph(Surv(eventtime, event) ~ strata(X1), 
                     data = d, y = TRUE, x = TRUE)

## compare models
resH0 <- influenceCoxTest(list(m.cox, mStrata.cox),
                 type = "survival", newdata = newdata, times = unique(d$eventtime))
resH0[1:10,]

#### Under H0 (CSC) ####
set.seed(1)
d <- sampleData(n, outcome = "competing.risk")
setkey(d, time)

## fit cox models
m.CSC <- CSC(Hist(time, event) ~ X1 + X2, data = d)
mStrata.CSC <- CSC(Hist(time, event) ~ strata(X1) + X2, data = d)

## compare models
resH0 <- influenceCoxTest(list(m.CSC, mStrata.CSC),
                 cause = 1, newdata = data.frame(X1=1,X2=1), times = unique(d$time))
resH0[1:10,]

}
