% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.matrix.R
\name{model.matrix.cph}
\alias{model.matrix.cph}
\title{Extract design matrix for cph objects}
\usage{
\method{model.matrix}{cph}(object, data, ...)
}
\arguments{
\item{object}{a cph object.}

\item{data}{a dataset.}

\item{...}{not used}
}
\description{
Extract design matrix for cph objects
}
