% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoxInfo.R
\name{model.matrix.phreg}
\alias{model.matrix.phreg}
\title{Extract design matrix for phreg objects}
\usage{
\method{model.matrix}{phreg}(object, data)
}
\arguments{
\item{object}{a phreg object.}

\item{data}{a dataset.}
}
\description{
Extract design matrix for phreg objects
}
\details{
mainly a copy paste of the begining of the \code{phreg} function.
}
