% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.CauseSpecificCox.R
\name{predict.CauseSpecificCox}
\alias{predict.CauseSpecificCox}
\title{Predicting absolute risk from cause-specific Cox models}
\usage{
\method{predict}{CauseSpecificCox}(object, newdata, times, cause, t0 = NA,
  colnames = TRUE, se = FALSE, ...)
}
\arguments{
\item{object}{The fitted cause specific Cox model}

\item{newdata}{A data frame containing the values of the variables
in the right hand side of 'coxph' for each patient.}

\item{times}{Vector of times at which to return the estimated
hazards/survival}

\item{cause}{Identifies the cause of interest among the competing events.}

\item{t0}{the starting time for the conditional survival.}

\item{colnames}{Logical. If \code{TRUE} name the columns of the matrix containing the absolute risk with times}

\item{se}{Logical. If \code{TRUE} add the standard errors corresponding to the output}

\item{...}{not used}
}
\value{
A data table containing the predictions for each patient (in rows) and each time (in columns).
}
\description{
Apply formula to combine two or more Cox models into absolute risk (cumulative incidence function)
}
\details{
Note: for Cox regression models with time varying
    covariates it does not make sense to use this function, because
    the predicted risk has to be a measurable function of the data
    available at the time origin.
}
\examples{

d <- SimCompRisk(1e2)
d$time <- round(d$time,1)
ttt <- sample(x = unique(sort(d$time)), size = 10) 

#### coxph function
CSC.fit <- CSC(Hist(time,event)~ X1+X2,data=d, method = "breslow")

predCSC <- predict(CSC.fit, newdata = d, cause = 2, times = ttt)

#### cph function
CSC.cph <- CSC(Hist(time,event)~ X1+X2,data=d, method = "breslow", fitter = "cph")

predcph <- predict(CSC.cph, newdata = d, cause = 2, times = ttt)

#### conditional survival
T0 <- 1
predCSC_afterT0 <- predict(CSC.fit, newdata = d, cause = 2, times = ttt, t0 = T0)

}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds
    tag@biostat.ku.dk
}

