% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{predictCox}
\alias{predictCox}
\title{Fast prediction of survival, hazard and cumulative hazard from Cox regression model 

Fast routine to get baseline hazards and subject specific hazards
as well as survival probabilities from a coxph or cph object}
\usage{
predictCox(object, newdata = NULL, times, centered = TRUE, maxtime = Inf,
  type = c("hazard", "cumHazard", "survival"), keep.strata = FALSE,
  keep.times = FALSE)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{newdata}{A data frame or table containing the values of the
predictor variables defining subject specific predictions. Should have
the same structure as the data set used to fit the \code{object}.}

\item{times}{Time points at which to evaluate the predictions.}

\item{centered}{If TRUE remove the centering factor used by \code{coxph}
in the linear predictor.}

\item{maxtime}{Baseline hazard will be computed for each event before maxtime}

\item{type}{One or several strings that match (either in lower or upper case or mixtures) one
or several of the strings \code{"hazard"},\code{"cumhazard"}, \code{"survival"}}

\item{keep.strata}{Logical. If \code{TRUE} add the strata in the newdata. Only if there any.}
}
\value{
a data table containing the time, the strata (if any), the hazard and the cumulative hazard.
}
\details{
Not working with time varying predictor variables or
    delayed entry.
}
\examples{
library(survival)

set.seed(10)
d <- SimSurv(1e2)
d$time <- round(d$time,1)
fit <- coxph(Surv(time,status)~X1 * X2,data=d, ties="breslow")
# table(duplicated(d$time))

predictCox(fit)
cbind(survival::basehaz(fit),baseHazRR(fit))
predictCox(fit, maxtime = 5)

# strata
fitS <- coxph(Surv(time,status)~strata(X1)+X2,data=d, ties="breslow")

predictCox(fitS)
predictCox(fitS, maxtime = 5)
}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

