% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCox.R
\name{predictCox}
\alias{predictCox}
\title{Fast prediction of survival, hazard and cumulative hazard from Cox regression model}
\usage{
predictCox(object, newdata = NULL, times, centered = TRUE, maxtime = NULL,
  type = c("hazard", "cumHazard", "survival"), keep.strata = FALSE,
  keep.times = FALSE, keep.lastEventTime = FALSE, se = FALSE)
}
\arguments{
\item{object}{The fitted Cox regression model object either
obtained with \code{coxph} (survival package) or \code{cph}
(rms package).}

\item{newdata}{A data frame or table containing the values of the
predictor variables defining subject specific predictions. Should have
the same structure as the data set used to fit the \code{object}.}

\item{times}{Time points at which to evaluate the predictions.}

\item{centered}{If TRUE remove the centering factor used by \code{coxph}
in the linear predictor.}

\item{maxtime}{Baseline hazard will be computed for each event before maxtime}

\item{type}{One or several strings that match (either in lower or upper case or mixtures) one
or several of the strings \code{"hazard"},\code{"cumhazard"}, \code{"survival"}}

\item{keep.strata}{Logical. If \code{TRUE} add the (newdata) strata to the output. Only if there any.}

\item{keep.times}{Logical. If \code{TRUE} add the evaluation times to the output.}

\item{keep.lastEventTime}{Logical. If \code{TRUE} add the time at which the last event occured for each strata to the output.}

\item{se}{Logical. If \code{TRUE} add the standard errors corresponding to the output. Experimental !!}
}
\value{
A list optionally containing the time, the strata (if any), the hazard, the
        cumulative hazard and survival probabilities.
}
\description{
Fast routine to get baseline hazards and subject specific hazards
as well as survival probabilities from a coxph or cph object
}
\details{
Not working with time varying predictor variables or
    delayed entry.
}
\examples{
library(survival)

set.seed(10)
d <- SimSurv(1e2)
nd <- SimSurv(10)
d$time <- round(d$time,1)
fit <- coxph(Surv(time,status)~X1 * X2,data=d, ties="breslow")
# table(duplicated(d$time))

predictCox(fit)
predictCox(fit, newdata=nd, times = 5)
cbind(survival::basehaz(fit),predictCox(fit,type="cumHazard"))
predictCox(fit, maxtime = 5,keep.times=TRUE)
predictCox(fit, maxtime = 1,keep.times=TRUE)

# one strata variable
fitS <- coxph(Surv(time,status)~strata(X1)+X2,data=d, ties="breslow")

predictCox(fitS)
predictCox(fitS, maxtime = 5)
predictCox(fitS, maxtime = 5, newdata=nd, times = 1)

# two strata variables
set.seed(1)
d$U=sample(letters[1:5],replace=TRUE,size=NROW(d))
d$V=sample(letters[4:10],replace=TRUE,size=NROW(d))
nd$U=sample(letters[1:5],replace=TRUE,size=NROW(nd))
nd$V=sample(letters[4:10],replace=TRUE,size=NROW(nd))
fit2S <- coxph(Surv(time,status)~X1+strata(U)+strata(V)+X2,data=d, ties="breslow")

cbind(survival::basehaz(fit2S),predictCox(fit2S,type="cumHazard"))
predictCox(fit2S)
predictCox(fitS, maxtime = 5)
predictCox(fitS, maxtime = 5,newdata=nd, times = 3)


}
\author{
Brice Ozenne broz@sund.ku.dk, Thomas A. Gerds tag@biostat.ku.dk
}

