% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictCoxPL.R
\name{predictCoxPL}
\alias{predictCoxPL}
\title{Computation of survival probabilities from Cox regression models using the product limit estimator.}
\usage{
predictCoxPL(object, newdata, times, type = c("cumhazard", "survival"),
  se = FALSE, band = FALSE, ...)
}
\description{
Same as predictCox except that the survival is estimated using the product limit estimator.
}
\examples{
library(survival)

set.seed(10)
d <- sampleData(40,outcome="survival")
nd <- sampleData(4,outcome="survival")
d$time <- round(d$time,1)
fit <- coxph(Surv(time,event)~X1 + strata(X2) + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)
predictCoxPL(fit, newdata = d, times = 1:5)
fit <- coxph(Surv(time,event)~X1 + X2 + X6,
             data=d, ties="breslow", x = TRUE, y = TRUE)
predictCoxPL(fit, newdata = d, times = 1:5)


#### Compare exp to product limit
set.seed(10)
A <- predictCoxPL(fit, newdata = d[1:5], times = 1:5, se = TRUE, band = TRUE, logTransform = FALSE)
set.seed(10)
B <- predictCox(fit, newdata = d[1:5], times = 1:5, se = TRUE, band = TRUE, logTransform = FALSE)

A$survival - B$survival
A$survival.lower - B$survival.lower
A$survival.upper - B$survival.upper
A$survival.lowerBand - B$survival.lowerBand
A$survival.upperBand - B$survival.upperBand
}
