% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRiskIID.R
\name{predictRiskIID}
\alias{predictRiskIID}
\alias{predictRiskIID.default}
\alias{predictRiskIID.glm}
\alias{predictRiskIID.coxph}
\alias{predictRiskIID.cph}
\alias{predictRiskIID.phreg}
\alias{predictRiskIID.CauseSpecificCox}
\title{Extract corrected i.i.d. decomposition}
\usage{
predictRiskIID(object, newdata, average.iid, factor, ...)

\method{predictRiskIID}{default}(object, newdata, average.iid, factor, ...)

\method{predictRiskIID}{glm}(object, newdata, average.iid, factor = NULL,
  ...)

\method{predictRiskIID}{coxph}(object, newdata, average.iid,
  factor = NULL, times, diag = FALSE, ...)

\method{predictRiskIID}{cph}(object, newdata, average.iid, factor = NULL,
  times, diag = FALSE, ...)

\method{predictRiskIID}{phreg}(object, newdata, average.iid,
  factor = NULL, times, diag = FALSE, ...)

\method{predictRiskIID}{CauseSpecificCox}(object, newdata, average.iid,
  factor = NULL, times, ...)
}
\arguments{
\item{object}{A fitted model from which to extract i.i.d. decomposition of the fitted values.}

\item{newdata}{A data frame containing predictor variable combinations.}

\item{average.iid}{Should the i.i.d. decomposition be averaged over individual observations in argument \code{newdata}?}

\item{factor}{When \code{average.iid} is TRUE, enables to perform weighted averages instead of averages.}

\item{...}{for compatibility with the generic method.}

\item{times}{A vector of times in the range of the response variable,
e.g. for which the cumulative incidences event probabilities are computed.
Disregarded when argument \code{object} is a logistic regression.}

\item{diag}{[logical] If \code{TRUE} only compute the survival for the i-th row in dataset at the i-th time. Only relevant for Cox models.}
}
\value{
\code{average.iid} is FALSE: an array containing
the influence of each individual from the training dataset at each time on each prediction.
\code{average.iid} is TRUE: a list where each element contains a "weighted" averaged i.i.d. decomposition.
}
\description{
Extract corrected i.i.d. decomposition8 from a gaussian linear model.
}
\details{
Argument \code{factor} must be a matrix with the same number of rows as argument \code{newdata}.
For given column, the i.i.d. decomposition is multiplied by the column (i.e. each individual contribution is weighted) before taking the average.
So the first element in the first column corresponds to the weight given to the first individual.
Therefore for each column, an averaged i.i.d. decomposition is returned.
Note that a one-column matrix with only 1 is equivalent to an average, i.e. not specifying the argument \code{factor}.
}
\seealso{
\code{\link{iidCox}} to obtain the iid decomposition for Cox models.
}
