% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ate.R
\name{print.ate}
\alias{print.ate}
\title{Print average treatment effects}
\usage{
\method{print}{ate}(x, digits = 3, ...)
}
\arguments{
\item{x}{object obtained with function \code{ate}}

\item{digits}{Number of digits}

\item{...}{passed to print}
}
\description{
Print average treatment effects
}
