% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ateRobust.R
\name{print.ateRobust}
\alias{print.ateRobust}
\title{Print Average Treatment Effect}
\usage{
\method{print}{ateRobust}(x, digits = 3, augment.cens = x$augment.cens,
  ...)
}
\arguments{
\item{x}{object obtained with the function \code{ateRobust}.}

\item{digits}{[integer, >0] indicating the number of decimal places.}

\item{augment.cens}{[logical] should the standard errors account for the augmentation term
in direction of the model for the censoring mechanism.
accounting for the uncertainty in the outcome and propensity score models be output}

\item{...}{Passed to print.}
}
\description{
Print average treatment effect.
}
