% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.IPA.R
\name{print.IPA}
\alias{print.IPA}
\title{Print IPA object}
\usage{
\method{print}{IPA}(x, digits = 2, ...)
}
\arguments{
\item{x}{Object obtained with \code{IPA}}

\item{digits}{Number of digits}

\item{...}{passed to print}
}
\description{
Print method for IPA
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
