% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Score.R
\name{print.Score}
\alias{print.Score}
\title{Print Score object}
\usage{
\method{print}{Score}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object obtained with \code{Score.list}}

\item{digits}{Number of digits}

\item{...}{passed to print}
}
\description{
Print method for risk prediction scores
}

