% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructData.R
\name{reconstructData}
\alias{reconstructData}
\title{Reconstruct the original dataset}
\usage{
reconstructData(object)
}
\arguments{
\item{object}{a coxph object.}
}
\description{
Reconstruct the original dataset from the elements stored in the coxph object
}
\author{
Brice Ozenne broz@sund.ku.dk and Thomas A. Gerds tag@biostat.ku.dk
}
