% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitStrataVar.R
\name{splitStrataVar}
\alias{splitStrataVar}
\title{Reconstruct each of the strata variables}
\usage{
splitStrataVar(object)
}
\arguments{
\item{object}{a coxph object.}
}
\description{
Reconstruct each of the strata variables from the strata variable stored in the coxph object.
}
\author{
Brice Ozenne broz@sund.ku.dk and Thomas A. Gerds tag@biostat.ku.dk
}
