% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize.R
\name{synthesize}
\alias{synthesize}
\title{Cooking and synthesizing survival data}
\usage{
synthesize(object, data, ...)
}
\arguments{
\item{...}{}
}
\value{
lava object
}
\description{
Synthesizing survival data
}
\details{
The simulation engine is: lava.
}
\examples{
# pbc data
library(survival)
library(lava)
u <- lvm()
distribution(u,~sex) <- binomial.lvm()
distribution(u,~age) <- normal.lvm()
distribution(u,~trt) <- binomial.lvm()
distribution(u,~logbili) <- normal.lvm()
distribution(u,~logprotime) <- normal.lvm()
u <-eventTime(u,time~min(time.cens=0,time.transplant=1,time.death=2), "status")
lava::regression(u,logbili~age+sex) <- 1
lava::regression(u,protimegrp1~age+sex+logbili) <- 1
lava::regression(u,protimegrp2~age+sex+logbili) <- 1
lava::regression(u,stage3~age+sex+protimegrp1+protimegrp2+logbili) <- 1
lava::regression(u,stage4~age+sex+protimegrp1+protimegrp2+logbili) <- 1
lava::regression(u,time.transplant~sex+age+logbili+protimegrp1+protimegrp2+stage3+stage4) <- 1
lava::regression(u,time.death~sex+age+logbili+protimegrp1+protimegrp2+stage3+stage4) <- 1
lava::regression(u,time.cens~1) <- 1
u <- categorical(u,~stage,labels=c("1/2","3","4"), K=3)
transform(u,protime~logprotime) <- function(x){exp(x)}
transform(u,protimegrp~protime) <- function(x){cut(c(unlist(x)), c(-Inf,10,11,Inf), labels=c("0","1","2"))}
u_synt <- synthesize(object=u, data=na.omit(pbc))
set.seed(8)
d <- sim(u_synt,n=1000)
fit_sim <- coxph(Surv(time,status==1)~age+sex+logbili,data=d)
fit_real <- coxph(Surv(time,status==1)~age+sex+logbili,data=pbc)
# compare estimated log-hazard ratios between simulated and real data 
cbind(coef(fit_sim),coef(fit_real))

data(pbc)
pbc <- na.omit(pbc[,c("time","status","sex","age","bili")])
pbc$logbili <- log(pbc$bili)
v_synt <- synthesize(object=Surv(time,status)~sex+age+logbili,data=pbc)
d <- sim(v_synt,1000)
fit_sim <- coxph(Surv(time,status==1)~age+sex+logbili,data=d)
fit_real <- coxph(Surv(time,status==1)~age+sex+logbili,data=pbc)
# compare estimated log-hazard ratios between simulated and real data
cbind(coef(fit_sim),coef(fit_real))

#
# Cancer data
#
data(cancer)
b <- lvm()
distribution(b,~rx) <- binomial.lvm()
distribution(b,~age) <- normal.lvm()
distribution(b,~resid.ds) <- binomial.lvm()
distribution(b,~ecog.ps) <- binomial.lvm()
lava::regression(b,time.death~age+rx+resid.ds) <- 1
b<-eventTime(b,futime~min(time.cens=0,time.death=1), "fustat")
b_synt <- synthesize(object = b, data = ovarian)
D <- sim(b_synt,1000)
fit_real <- coxph(Surv(futime,fustat)~age+rx+resid.ds, data=ovarian)
fit_sim <- coxph(Surv(futime,fustat)~age+rx+resid.ds, data=D)
cbind(coef(fit_sim),coef(fit_real))
w_synt <- synthesize(object=Surv(futime,fustat)~age+rx+resid.ds, data=ovarian)
D <- sim(w_synt,1000)
fit_sim <- coxph(Surv(futime,fustat==1)~age+rx+resid.ds,data=D)
fit_real <- coxph(Surv(futime,fustat==1)~age+rx+resid.ds,data=ovarian)
# compare estimated log-hazard ratios between simulated and real data
cbind(coef(fit_sim),coef(fit_real))


}
\seealso{
lvm
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
