% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transformIID}
\alias{transformIID}
\title{Compute Influence Functions after Transformation}
\usage{
transformIID(estimate, iid, type)
}
\arguments{
\item{estimate}{[numeric matrix] the estimate value before transformation.}

\item{iid}{[numeric array] the standard error before transformation.}

\item{type}{[character] the transforamtion.
Can be \code{"log"}, \code{"loglog"}, \code{"cloglog"}, or \code{"atanh"} (Fisher transform).}
}
\description{
Compute influence functions after transformation
based on the influence function before transformation.
}
\details{
Use a delta method to find the standard error after transformation. \cr \cr

The iid decomposition must contain have dimension [n.obs,time,n.prediction] and estimate [n.prediction,time].
}
