% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglm.R
\name{weights.wglm}
\alias{weights.wglm}
\title{Extract IPCW Weights}
\usage{
\method{weights}{wglm}(object, times = NULL, prefix = "t", simplify = TRUE, ...)
}
\arguments{
\item{object}{a wglm object.}

\item{times}{[numeric vector] time points at which the weights should be output.}

\item{prefix}{[character] used to name the columns. Can be \code{NA} to keep the original names.}

\item{simplify}{[logical] should the ouput be converted to a vector when only one timepoint is requested. Otherwise will always return a matrix.}

\item{...}{Not used.}
}
\description{
Extract IPCW weights of IPCW logistic regressions.
}
